@@ -1,8 +1,7 @@
 {
   "name": "mirakurun",
-  "preferGlobal": true,
   "description": "DVR Tuner Server for Japanese TV.",
-  "version": "4.0.0-beta.2",
+  "version": "4.0.0-beta.12",
   "homepage": "https://github.com/Chinachu/Mirakurun",
   "keywords": [
     "mirakurun",
@@ -31,19 +30,21 @@
   },
   "license": "Apache-2.0",
   "bin": {
+    "mirakurun": "bin/cli.sh",
     "mirakurun-epgdump": "bin/epgdump.js"
   },
   "main": "lib/client.js",
   "scripts": {
-    "start": "node -r source-map-support/register lib/server.js",
-    "debug": "node -r source-map-support/register --inspect=0.0.0.0:9229 lib/server.js",
+    "start": "node --max-semi-space-size=64 -r source-map-support/register lib/server.js ",
+    "debug": "node --max-semi-space-size=64 -r source-map-support/register --inspect=0.0.0.0:9229 lib/server.js",
     "build": "npm run build:server && npm run build:webpack",
     "build:server": "tslint --project . && tsc --declaration",
     "build:webpack": "webpack",
     "watch": "concurrently -n w: npm:watch:*",
     "watch:server": "tsc -w --declaration",
     "watch:webpack": "webpack -w",
-    "test": "tslint --project . && mocha --exit test/*.spec.js",
+    "test": "node --test --test-reporter=spec -r source-map-support/register",
+    "test:watch": "node --test --watch --test-reporter=tap -r source-map-support/register",
     "clean": "rimraf lib/*",
     "prepublishOnly": "npm run clean && npm run build",
     "postinstall": "node bin/postinstall.js",
@@ -51,7 +52,9 @@
     "docker:run-setup": "docker compose -f docker/docker-compose.yml run --rm -e SETUP=true --service-ports mirakurun",
     "docker:run": "docker compose -f docker/docker-compose.yml run --rm --service-ports mirakurun",
     "docker:up": "docker compose -f docker/docker-compose.yml up -d",
+    "docker:down": "docker compose -f docker/docker-compose.yml down",
     "docker:logs": "docker compose -f docker/docker-compose.yml logs -f",
+    "docker:bash": "docker compose -f docker/docker-compose.yml exec mirakurun bash",
     "docker:debug": "docker compose -f docker/docker-compose.yml run --rm --service-ports -e DEBUG=true mirakurun"
   },
   "directories": {
@@ -66,7 +69,6 @@
     "eventemitter3": "4.0.7",
     "express": "4.21.2",
     "express-openapi": "8.0.0",
-    "glob": "7.2.3",
     "ip-num": "1.3.4",
     "js-yaml": "4.1.0",
     "jsonrpc2-ws": "1.0.0-beta23",
@@ -74,40 +76,40 @@
     "morgan": "1.10.0",
     "openapi-types": "7.2.3",
     "promise-queue": "2.2.5",
-    "rfdc": "1.3.0",
+    "redoc": "2.4.0",
+    "redoc-try": "1.4.10",
+    "rfdc": "1.4.1",
     "semver": "7.7.1",
     "sift": "15.1.3",
     "source-map-support": "0.5.21",
-    "swagger-ui-dist": "4.6.1",
-    "tail": "2.2.4"
+    "yieldable-json": "^2.1.0"
   },
   "devDependencies": {
-    "@fluentui/react": "8.49.6",
+    "@fluentui/react": "^8.122.11",
     "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
+    "@types/express": "^4.17.21",
     "@types/js-yaml": "^4.0.2",
+    "@types/mime": "^3.0.4",
     "@types/morgan": "^1.9.3",
     "@types/node": "18",
     "@types/promise-queue": "^2.2.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "17.0.13",
-    "@types/swagger-ui-dist": "^3.30.1",
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
     "@types/ws": "^7.4.7",
+    "@types/yieldable-json": "^2.0.2",
     "buffer": "^6.0.3",
     "concurrently": "^7.0.0",
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "5.2.7",
-    "mocha": "11.1.0",
     "process": "^0.11.10",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "rimraf": "^3.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "rimraf": "^6.0.1",
     "style-loader": "^2.0.0",
-    "ts-loader": "9.2.7",
+    "ts-loader": "9.5.2",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "5.7",
-    "url": "^0.11.0",
     "webpack": "^5.77.0",
     "webpack-cli": "^4.9.2"
   },
