@@ -42,16 +42,10 @@ export interface Channel {
     type: ChannelType;
     channel: string;
     name?: string;
-    satellite?: string;
-    space?: number;
-    freq?: number;
-    polarity?: "H" | "V";
     services?: Service[];
 }
 
 export type ChannelType = "GR" | "BS" | "CS" | "SKY" | "BS4K";
-
-export type ChannelScanMode = "Channel" | "Service";
 
 export interface Service {
     id: ServiceItemId;
@@ -85,24 +79,7 @@ export interface Program {
         streamContent: number;
         componentType: number;
     }
-    audios?: {
-        /** component_type
-         * - 0x01 - 1/0 mode (single-mono)
-         * - 0x02 - 1/0 + 1/0 mode (dual-mono)
-         * - 0x03 - 2/0 mode (stereo)
-         * - 0x07 - 3/1 mode
-         * - 0x08 - 3/2 mode
-         * - 0x09 - 3/2 + LFE mode
-         */
-        componentType: number;
-        componentTag: number;
-        isMain: boolean;
-        samplingRate: ProgramAudioSamplingRate;
-        /** ISO_639_language_code, ISO_639_language_code_2
-         * - this `#length` will `2` if dual-mono multi-lingual.
-         */
-        langs: ProgramAudioLanguageCode[];
-    }[]
+    audios?: ProgramAudio[];
 
     series?: ProgramSeries;
 
@@ -122,16 +99,44 @@ export interface ProgramGenre {
 
 export type ProgramVideoType = "mpeg2" | "h.264" | "h.265";
 
-export type ProgramVideoResolution = "240p" | "480i" | "480p" | "720p" | "1080i" | "2160p" | "4320p";
+export type ProgramVideoResolution = (
+    "240p" |
+    "480i" |
+    "480p" |
+    "720p" |
+    "1080i" |
+    "1080p" |
+    "2160p" |
+    "4320p"
+);
 
-export enum ProgramAudioSamplingRate {
-    "16kHz" = 16000,
-    "22.05kHz" = 22050,
-    "24kHz" = 24000,
-    "32kHz" = 32000,
-    "44.1kHz" = 44100,
-    "48kHz" = 48000
+export interface ProgramAudio {
+    /** component_type
+     * - 0x01 - 1/0 mode (single-mono)
+     * - 0x02 - 1/0 + 1/0 mode (dual-mono)
+     * - 0x03 - 2/0 mode (stereo)
+     * - 0x07 - 3/1 mode
+     * - 0x08 - 3/2 mode
+     * - 0x09 - 3/2 + LFE mode
+     */
+    componentType: number;
+    componentTag: number;
+    isMain: boolean;
+    samplingRate: ProgramAudioSamplingRate;
+    /** ISO_639_language_code, ISO_639_language_code_2
+     * - this `#length` will `2` if dual-mono multi-lingual.
+     */
+    langs: ProgramAudioLanguageCode[];
 }
+
+export type ProgramAudioSamplingRate = (
+    16000 |
+    22050 |
+    24000 |
+    32000 |
+    44100 |
+    48000
+);
 
 export type ProgramAudioLanguageCode = (
     "jpn" |
@@ -211,10 +216,10 @@ export interface TunerProcess {
     pid: number;
 }
 
-export interface Event {
+export interface Event<T = any> {
     resource: EventResource;
     type: EventType;
-    data: any;
+    data: T;
     time: UnixtimeMS;
 }
 
@@ -241,15 +246,19 @@ export interface ConfigServer {
     allowIPv6CidrRanges?: string[];
     useTSId?: boolean;
     useStreamId?: boolean;
+    allowOrigins: string[];
+    allowPNA: boolean;
+    tsplayEndpoint: string;
 }
 
-export enum LogLevel {
-    "FATAL" = -1,
-    "ERROR" = 0,
-    "WARN" = 1,
-    "INFO" = 2,
-    "DEBUG" = 3
-}
+/**
+ * FATAL: -1
+ * ERROR: 0
+ * WARN: 1
+ * INFO: 2
+ * DEBUG: 3
+ */
+export type LogLevel = -1 | 0 | 1 | 2 | 3;
 
 export type ConfigTuners = ConfigTunersItem[];
 
@@ -282,14 +291,69 @@ export interface ConfigChannelsItem {
     type: ChannelType;
     /** passed to tuning command */
     channel: string;
-    satellite?: string;
     serviceId?: number;
-    space?: number;
-    freq?: number;
-    polarity?: "H" | "V";
+    /** TSMF (MPEG-TS Multi Frame) relative TS number config for CATV */
     tsmfRelTs?: number;
+    /**
+     * passed to tuning command variables.
+     * @example { "freq": 123456, "polarity": "H", "space": 6, "extra-args": "..." }
+     */
+    commandVars?: Record<string, string | number>;
     isDisabled?: boolean;
+    /** @deprecated typo of "satellite". */
+    readonly satelite?: string;
+    /** @deprecated from 4.0.0, use `commandVars` instead. */
+    readonly satellite?: string;
+    /** @deprecated from 4.0.0, use `commandVars` instead. */
+    readonly space?: number;
+    /** @deprecated from 4.0.0, use `commandVars` instead. */
+    readonly freq?: number;
+    /** @deprecated from 4.0.0, use `commandVars` instead. */
+    readonly polarity?: "H" | "V";
 }
+
+export interface ChannelScanStatus {
+    isScanning: boolean;
+    status: ChannelScanPhase;
+    type?: ChannelType;
+    dryRun?: boolean;
+    progress?: number;
+    currentChannel?: string;
+    scanLog?: string[];
+    newCount?: number;
+    takeoverCount?: number;
+    result?: ConfigChannelsItem[];
+    startTime?: number;
+    updateTime?: number;
+}
+
+export type ChannelScanMode = "Channel" | "Service";
+
+export type ChannelScanPhase = (
+    "not_started" |
+    "scanning" |
+    "completed" |
+    "cancelled" |
+    "error"
+);
+
+export type ChannelScanStep = (
+    "started" |
+    "scanning_channel" |
+    "takeover" |
+    "skipped" |
+    "services_found" |
+    "channels_found" |
+    "error"
+);
+
+export type ChannelScanResultType = (
+    "summary" |
+    "summary_new" |
+    "summary_takeover" |
+    "restart_required" |
+    "final_result"
+);
 
 export interface Version {
     current: string;
