@@ -13,20 +13,18 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-import * as common from "./common";
 import * as log from "./log";
+import * as apid from "../../api";
 import _ from "./_";
 import status from "./status";
 import queue from "./queue";
 import ChannelItem from "./ChannelItem";
 
-export default class Channel {
-
+export class Channel {
     private _items: ChannelItem[] = [];
     private _epgGatheringInterval: number = _.config.server.epgGatheringInterval || 1000 * 60 * 30; // 30 mins
 
     constructor() {
-
         this._load();
 
         if (_.config.server.disableEITParsing !== true) {
@@ -39,14 +37,12 @@ export default class Channel {
     }
 
     add(item: ChannelItem): void {
-
         if (this.get(item.type, item.channel) === null) {
             this._items.push(item);
         }
     }
 
-    get(type: common.ChannelType, channel: string): ChannelItem {
-
+    get(type: apid.ChannelType, channel: string): ChannelItem {
         const l = this._items.length;
         for (let i = 0; i < l; i++) {
             if (this._items[i].channel === channel && this._items[i].type === type) {
@@ -57,8 +53,7 @@ export default class Channel {
         return null;
     }
 
-    findByType(type: common.ChannelType): ChannelItem[] {
-
+    findByType(type: apid.ChannelType): ChannelItem[] {
         const items = [];
 
         const l = this._items.length;
@@ -72,13 +67,11 @@ export default class Channel {
     }
 
     private _load(): void {
-
         log.debug("loading channels...");
 
         const channels = _.config.channels;
 
         channels.forEach((channel, i) => {
-
             if (typeof channel.name !== "string") {
                 log.error("invalid type of property `name` in channel#%d configuration", i);
                 return;
@@ -94,31 +87,6 @@ export default class Channel {
                 return;
             }
 
-            if (channel.satelite && !channel.satellite) {
-                log.warn("renaming deprecated property name `satelite` to `satellite` in channel#%d (%s) configuration", i, channel.name);
-                (<any> channel).satellite = channel.satelite;
-            }
-
-            if (channel.satellite && typeof channel.satellite !== "string") {
-                log.error("invalid type of property `satellite` in channel#%d (%s) configuration", i, channel.name);
-                return;
-            }
-
-            if (channel.space && typeof channel.space !== "number") {
-                log.error("invalid type of property `space` in channel#%d (%s) configuration", i, channel.name);
-                return;
-            }
-
-            if (channel.freq !== undefined && typeof channel.freq !== "number") {
-                log.error("invalid type of property `freq` in channel#%d (%s) configuration", i, channel.name);
-                return;
-            }
-
-            if (channel.polarity && channel.polarity !== "H" && channel.polarity !== "V") {
-                log.error("invalid type of property `polarity` in channel#%d (%s) configuration", i, channel.name);
-                return;
-            }
-
             if (channel.serviceId && typeof channel.serviceId !== "number") {
                 log.error("invalid type of property `serviceId` in channel#%d (%s) configuration", i, channel.name);
                 return;
@@ -127,6 +95,48 @@ export default class Channel {
             if (channel.tsmfRelTs && typeof channel.tsmfRelTs !== "number") {
                 log.error("invalid type of property `tsmfRelTs` in channel#%d (%s) configuration", i, channel.name);
                 return;
+            }
+
+            if (channel.commandVars && typeof channel.commandVars !== "object") {
+                log.error("invalid type of property `commandVars` in channel#%d (%s) configuration", i, channel.name);
+                return;
+            }
+            if (!channel.commandVars) {
+                channel.commandVars = {};
+            }
+            if (channel.satelite && !channel.satellite) {
+                log.warn("renaming deprecated property name `satelite` to `satellite` in channel#%d (%s) configuration", i, channel.name);
+                (<any> channel).satellite = channel.satelite;
+            }
+            if (channel.satellite) {
+                // deprecated but not planned to remove (soft migration)
+                if (!channel.commandVars.satellite) {
+                    channel.commandVars.satellite = channel.satellite;
+                }
+            }
+            if (channel.space) {
+                // deprecated but not planned to remove (soft migration)
+                if (!channel.commandVars.space) {
+                    channel.commandVars.space = channel.space;
+                }
+            }
+            if (channel.freq) {
+                // deprecated but not planned to remove (soft migration)
+                if (!channel.commandVars.freq) {
+                    channel.commandVars.freq = channel.freq;
+                }
+            }
+            if (channel.polarity) {
+                // deprecated but not planned to remove (soft migration)
+                if (!channel.commandVars.polarity) {
+                    channel.commandVars.polarity = channel.polarity;
+                }
+            }
+            for (const key in channel.commandVars) {
+                if (typeof channel.commandVars[key] !== "number" && typeof channel.commandVars[key] !== "string") {
+                    log.error("invalid type of property `commandVars.%s` in channel#%d (%s) configuration", key, i, channel.name);
+                    delete channel.commandVars[key];
+                }
             }
 
             if (channel.isDisabled === true) {
@@ -152,9 +162,7 @@ export default class Channel {
     }
 
     private _epgGatherer(): void {
-
         queue.add(async () => {
-
             const networkIds = [...new Set(_.service.items.map(item => item.networkId))];
 
             networkIds.forEach(networkId => {
@@ -170,7 +178,6 @@ export default class Channel {
                 const service = services[0];
 
                 queue.add(async () => {
-
                     if (service.epgReady === true) {
                         const now = Date.now();
                         if (now - service.epgUpdatedAt < this._epgGatheringInterval) {
@@ -272,3 +279,5 @@ export default class Channel {
         });
     }
 }
+
+export default Channel;
