@@ -20,30 +20,23 @@ import * as db from "./db";
 import * as log from "./log";
 import * as common from "./common";
 import * as config from "./config";
+import * as apid from "../../api";
 import ServiceItem from "./ServiceItem";
 import { StreamFilter } from "./StreamFilter";
 
 export default class ChannelItem {
+    readonly name: string;
+    readonly type: apid.ChannelType;
+    readonly channel: string;
+    readonly tsmfRelTs: number;
+    readonly commandVars: apid.ConfigChannelsItem["commandVars"];
 
-    private _name: string;
-    private _type: common.ChannelType;
-    private _channel: string;
-    private _satellite: string;
-    private _space: number;
-    private _freq: number;
-    private _polarity: "H" | "V";
-    private _tsmfRelTs: number;
-
-    constructor(config: config.Channel) {
-
-        this._name = config.name;
-        this._type = config.type;
-        this._channel = config.channel;
-        this._satellite = config.satellite;
-        this._space = config.space;
-        this._freq = config.freq;
-        this._polarity = config.polarity;
-        this._tsmfRelTs = config.tsmfRelTs;
+    constructor(config: apid.ConfigChannelsItem) {
+        this.name = config.name;
+        this.type = config.type;
+        this.channel = config.channel;
+        this.tsmfRelTs = config.tsmfRelTs;
+        this.commandVars = config.commandVars;
 
         if (config.serviceId) {
             this.addService(config.serviceId);
@@ -58,53 +51,7 @@ export default class ChannelItem {
         }, 3000);
     }
 
-    get name(): string {
-        return this._name;
-    }
-
-    get type(): common.ChannelType {
-        return this._type;
-    }
-
-    get channel(): string {
-        return this._channel;
-    }
-
-    get satellite(): string {
-        return this._satellite;
-    }
-
-    get space(): number {
-        return this._space;
-    }
-
-    get freq(): number {
-        return this._freq;
-    }
-
-    get polarity(): "H" | "V" {
-        return this._polarity;
-    }
-
-    get tsmfRelTs(): number {
-        return this._tsmfRelTs;
-    }
-
-    toJSON(): config.Channel {
-        return {
-            type: this._type,
-            channel: this._channel,
-            name: this._name,
-            satellite: this._satellite,
-            space: this._space,
-            freq: this._freq,
-            polarity: this._polarity,
-            tsmfRelTs: this._tsmfRelTs
-        };
-    }
-
     addService(serviceId: number): void {
-
         if (!_.service) {
             process.nextTick(() => this.addService(serviceId));
             return;
@@ -114,17 +61,16 @@ export default class ChannelItem {
             return;
         }
 
-        log.debug("ChannelItem#'%s' serviceId=%d check has queued", this._name, serviceId);
+        log.debug("ChannelItem#'%s' serviceId=%d check has queued", this.name, serviceId);
 
         queue.add(async () => {
-
-            log.info("ChannelItem#'%s' serviceId=%d check has started", this._name, serviceId);
+            log.info("ChannelItem#'%s' serviceId=%d check has started", this.name, serviceId);
 
             let services;
             try {
                 services = await _.tuner.getServices(this);
             } catch (e) {
-                log.warn("ChannelItem#'%s' serviceId=%d check has failed [%s]", this._name, serviceId, e);
+                log.warn("ChannelItem#'%s' serviceId=%d check has failed [%s]", this.name, serviceId, e);
 
                 setTimeout(() => this.addService(serviceId), 180000);
                 return;
@@ -132,13 +78,13 @@ export default class ChannelItem {
 
             const service = services.find(service => service.serviceId === serviceId);
             if (!service) {
-                log.warn("ChannelItem#'%s' serviceId=%d check has failed [no service]", this._name, serviceId);
+                log.warn("ChannelItem#'%s' serviceId=%d check has failed [no service]", this.name, serviceId);
 
                 setTimeout(() => this.addService(serviceId), 3600000);
                 return;
             }
 
-            log.debug("ChannelItem#'%s' serviceId=%d: %s", this._name, serviceId, JSON.stringify(service, null, "  "));
+            log.debug("ChannelItem#'%s' serviceId=%d: %s", this.name, serviceId, JSON.stringify(service, null, "  "));
 
             _.service.add(
                 new ServiceItem(this, service.networkId, service.serviceId, service.name, service.type, service.logoId)
@@ -155,28 +101,25 @@ export default class ChannelItem {
     }
 
     serviceScan(add: boolean): void {
-
-        log.debug("ChannelItem#'%s' service scan has queued", this._name);
+        log.debug("ChannelItem#'%s' service scan has queued", this.name);
 
         queue.add(async () => {
+            log.info("ChannelItem#'%s' service scan has started", this.name);
 
-            log.info("ChannelItem#'%s' service scan has started", this._name);
-
-            let services: db.Service[];
+            let services: apid.Service[];
             try {
                 const r = await _.tuner.getServices(this);
                 services = r.services;
             } catch (e) {
-                log.warn("ChannelItem#'%s' service scan has failed [%s]", this._name, e);
+                log.warn("ChannelItem#'%s' service scan has failed [%s]", this.name, e);
 
                 setTimeout(() => this.serviceScan(add), add ? 180000 : 3600000);
                 return;
             }
 
-            log.debug("ChannelItem#'%s' services: %s", this._name, JSON.stringify(services, null, "  "));
+            log.debug("ChannelItem#'%s' services: %s", this.name, JSON.stringify(services, null, "  "));
 
             services.forEach(service => {
-
                 const item = _.service.get(service.networkId, service.serviceId);
                 if (item !== null) {
                     item.name = service.name;
@@ -200,7 +143,7 @@ export default class ChannelItem {
                 }
             });
 
-            log.info("ChannelItem#'%s' service scan has finished", this._name);
+            log.info("ChannelItem#'%s' service scan has finished", this.name);
         });
     }
 }
