@@ -15,7 +15,7 @@
 */
 import { getProgramItemId } from "./Program";
 import { getTimeFromMJD, getTimeFromBCD24 } from "./common";
-import * as db from "./db";
+import * as apid from "../../api";
 import _ from "./_";
 import { TsChar } from "@chinachu/aribts";
 import { EIT } from "@chinachu/aribts/lib/table/eit";
@@ -117,14 +117,14 @@ interface EventState {
     };
     audio: {
         version: VersionRecord<VersionRecord>; // basic
-        _audios: { [componentTag: number]: db.ProgramAudio };
+        _audios: { [componentTag: number]: apid.ProgramAudio };
     };
     series: {
         version: VersionRecord; // basic
     };
     group: {
         version: VersionRecord<VersionRecord>; // basic
-        _groups: db.ProgramRelatedItem[][];
+        _groups: apid.ProgramRelatedItem[][];
     };
 
     present?: true;
@@ -133,11 +133,9 @@ interface EventState {
 
 // forked from rndomhack/node-aribts/blob/1e7ef94bba3d6ac26aec764bf24dde2c2852bfcb/lib/epg.js
 export default class EPG {
-
     private _epg: { [networkId: number]: { [serviceId: number]: { [eventId: number]: EventState } } } = {};
 
     write(eit: EIT) {
-
         if (!this._epg) {
             return;
         }
@@ -319,8 +317,8 @@ export default class EPG {
 
                         _.program.set(state.programId, {
                             video: {
-                                type: <db.ProgramVideoType> STREAM_CONTENT[d.stream_content] || null,
-                                resolution: <db.ProgramVideoResolution> COMPONENT_TYPE[d.component_type] || null,
+                                type: <apid.ProgramVideoType> STREAM_CONTENT[d.stream_content] || null,
+                                resolution: <apid.ProgramVideoResolution> COMPONENT_TYPE[d.component_type] || null,
 
                                 streamContent: d.stream_content,
                                 componentType: d.component_type
@@ -420,7 +418,6 @@ export default class EPG {
 }
 
 function isOutOfDate(eit: EIT, versionRecord: VersionRecord): boolean {
-
     if (
         (versionRecord[0x4E] !== undefined && eit.table_id !== 0x4E) ||
         (versionRecord[0x4F] !== undefined && eit.table_id !== 0x4E && eit.table_id !== 0x4F)
@@ -432,7 +429,6 @@ function isOutOfDate(eit: EIT, versionRecord: VersionRecord): boolean {
 }
 
 function isOutOfDateLv2(eit: EIT, versionRecord: VersionRecord<VersionRecord>, lv2: number): boolean {
-
     if (
         (versionRecord[0x4E] !== undefined && eit.table_id !== 0x4E) ||
         (versionRecord[0x4F] !== undefined && eit.table_id !== 0x4E && eit.table_id !== 0x4F)
@@ -446,7 +442,7 @@ function isOutOfDateLv2(eit: EIT, versionRecord: VersionRecord<VersionRecord>, l
     return versionRecord[eit.table_id][lv2] !== eit.version_number;
 }
 
-function getGenre(content: any): db.ProgramGenre {
+function getGenre(content: any): apid.ProgramGenre {
     return {
         lv1: content.content_nibble_level_1,
         lv2: content.content_nibble_level_2,
@@ -455,16 +451,16 @@ function getGenre(content: any): db.ProgramGenre {
     };
 }
 
-function getLangCode(buffer: Buffer): db.ProgramAudioLanguageCode {
+function getLangCode(buffer: Buffer): apid.ProgramAudioLanguageCode {
     for (const code in ISO_639_LANG_CODE) {
         if (ISO_639_LANG_CODE[code].compare(buffer) === 0) {
-            return code as db.ProgramAudioLanguageCode;
+            return code as apid.ProgramAudioLanguageCode;
         }
     }
     return "etc";
 }
 
-function getRelatedProgramItem(event: any): db.ProgramRelatedItem {
+function getRelatedProgramItem(event: any): apid.ProgramRelatedItem {
     return {
         type: (
             this.group_type === 1 ? "shared" :
