@@ -16,13 +16,12 @@
 import * as stream from "stream";
 import * as common from "./common";
 import _ from "./_";
-import * as db from "./db";
+import * as apid from "../../api";
 import Event from "./Event";
 import ChannelItem from "./ChannelItem";
 import { StreamFilter } from "./StreamFilter";
 
 export default class ServiceItem {
-
     static getId(networkId: number, serviceId: number): number {
         return parseInt(networkId + (serviceId / 100000).toFixed(5).slice(2), 10);
     }
@@ -60,7 +59,6 @@ export default class ServiceItem {
     }
 
     set name(name: string) {
-
         if (this._name !== name) {
             this._name = name;
 
@@ -74,7 +72,6 @@ export default class ServiceItem {
     }
 
     set type(type: number) {
-
         if (this._type !== type) {
             this._type = type;
 
@@ -88,7 +85,6 @@ export default class ServiceItem {
     }
 
     set logoId(logoId: number) {
-
         if (this._logoId !== logoId) {
             this._logoId = logoId;
 
@@ -102,7 +98,6 @@ export default class ServiceItem {
     }
 
     set remoteControlKeyId(id: number) {
-
         if (this._remoteControlKeyId !== id) {
             this._remoteControlKeyId = id;
 
@@ -116,7 +111,6 @@ export default class ServiceItem {
     }
 
     set epgReady(epgReady: boolean) {
-
         if (this._epgReady !== epgReady) {
             this._epgReady = epgReady;
 
@@ -130,7 +124,6 @@ export default class ServiceItem {
     }
 
     set epgUpdatedAt(time: number) {
-
         if (this._epgUpdatedAt !== time) {
             this._epgUpdatedAt = time;
 
@@ -143,9 +136,8 @@ export default class ServiceItem {
         return this._channel;
     }
 
-    export(): db.Service {
-
-        const ret: db.Service = {
+    export(): apid.Service {
+        const ret: apid.Service = {
             id: this._id,
             serviceId: this._serviceId,
             networkId: this._networkId,
@@ -169,7 +161,6 @@ export default class ServiceItem {
     }
 
     getOrder(): number {
-
         let order: string;
 
         switch (this._channel.type) {
