@@ -16,9 +16,9 @@
 import { Writable } from "stream";
 import * as common from "./common";
 import * as log from "./log";
-import * as db from "./db";
+import * as apid from "../../api";
 import _ from "./_";
-import TunerDevice from "./TunerDevice";
+import TunerDevice, { TunerDeviceStatus } from "./TunerDevice";
 import ChannelItem from "./ChannelItem";
 import ServiceItem from "./ServiceItem";
 import TSFilter from "./TSFilter";
@@ -26,20 +26,18 @@ import TSDecoder from "./TSDecoder";
 import TLVFilter from "./TLVFilter";
 import { StreamFilter } from "./StreamFilter";
 
-export default class Tuner {
-
+export class Tuner {
     private _devices: TunerDevice[] = [];
 
     constructor() {
         this._load();
     }
 
-    get devices(): TunerDevice[] {
-        return this._devices;
+    get devices(): TunerDeviceStatus[] {
+        return this._devices.map(device => device.toJSON());
     }
 
     get(index: number): TunerDevice {
-
         const l = this._devices.length;
         for (let i = 0; i < l; i++) {
             if (this._devices[i].index === index) {
@@ -50,8 +48,7 @@ export default class Tuner {
         return null;
     }
 
-    typeExists(type: common.ChannelType): boolean {
-
+    typeExists(type: apid.ChannelType): boolean {
         const l = this._devices.length;
         for (let i = 0; i < l; i++) {
             if (this._devices[i].config.types.includes(type) === true) {
@@ -63,7 +60,6 @@ export default class Tuner {
     }
 
     initChannelStream(channel: ChannelItem, userReq: common.UserRequest, output: Writable): Promise<StreamFilter> {
-
         let networkId: number;
 
         const services = channel.getServices();
@@ -94,7 +90,7 @@ export default class Tuner {
         }, output);
     }
 
-    initProgramStream(program: db.Program, userReq: common.UserRequest, output: Writable): Promise<StreamFilter> {
+    initProgramStream(program: apid.Program, userReq: common.UserRequest, output: Writable): Promise<StreamFilter> {
 
         return this._initTS({
             ...userReq,
@@ -109,7 +105,6 @@ export default class Tuner {
     }
 
     async getEPG(channel: ChannelItem, time?: number): Promise<void> {
-
         let timeout: NodeJS.Timeout;
         if (!time) {
             time = _.config.server.epgRetrievalTime || 1000 * 60 * 10;
@@ -153,7 +148,7 @@ export default class Tuner {
         });
     }
 
-    async getServices(channel: ChannelItem): Promise<{ services: db.Service[], channels: db.Channel[] }> {
+    async getServices(channel: ChannelItem, user: Partial<common.User> = {}): Promise<{ services: apid.Service[], channels: apid.Channel[] }> {
 
         const tsFilter = await this._initTS({
             id: "Mirakurun:getServices()",
@@ -163,17 +158,18 @@ export default class Tuner {
                 channel,
                 parseNIT: true,
                 parseSDT: true
-            }
+            },
+            ...user
         });
-        return new Promise<{ services: db.Service[], channels: db.Channel[] }>((resolve, reject) => {
+        return new Promise<{ services: apid.Service[], channels: apid.Channel[] }>((resolve, reject) => {
 
             let network = {
                 networkId: -1,
                 areaCode: -1,
                 remoteControlKeyId: -1
             };
-            let services: db.Service[] = null;
-            let networkStreams: db.Channel[] = [];
+            let services: apid.Service[] = null;
+            let networkStreams: apid.Channel[] = [];
 
             setTimeout(() => tsFilter.close(), 20000);
 
@@ -199,7 +195,6 @@ export default class Tuner {
             ]).then(() => tsFilter.close());
 
             tsFilter.once("close", () => {
-
                 tsFilter.removeAllListeners("network");
                 tsFilter.removeAllListeners("services");
 
@@ -221,13 +216,11 @@ export default class Tuner {
     }
 
     private _load(): this {
-
         log.debug("loading tuners...");
 
         const tuners = _.config.tuners;
 
         tuners.forEach((tuner, i) => {
-
             if (!tuner.name || !tuner.types || (!tuner.remoteMirakurunHost && !tuner.command)) {
                 log.error("missing required property in tuner#%s configuration", i);
                 return;
@@ -284,7 +277,6 @@ export default class Tuner {
     }
 
     private _initTS(user: common.User, dest?: Writable): Promise<StreamFilter> {
-
         return new Promise<StreamFilter>((resolve, reject) => {
 
             const setting = user.streamSetting;
@@ -299,7 +291,6 @@ export default class Tuner {
             const length = devices.length;
 
             function find() {
-
                 let device: TunerDevice = null;
 
                 // 1. join to existing
@@ -436,8 +427,7 @@ export default class Tuner {
         });
     }
 
-    private _getDevicesByType(type: common.ChannelType): TunerDevice[] {
-
+    private _getDevicesByType(type: apid.ChannelType): TunerDevice[] {
         const devices = [];
 
         const l = this._devices.length;
@@ -450,3 +440,5 @@ export default class Tuner {
         return devices;
     }
 }
+
+export default Tuner;
