@@ -31,10 +31,10 @@ interface User extends common.User {
     _stream?: StreamFilter;
 }
 
-interface Status {
+export interface TunerDeviceStatus {
     readonly index: number;
     readonly name: string;
-    readonly types: common.ChannelType[];
+    readonly types: apid.ChannelType[];
     readonly command: string;
     readonly pid: number;
     readonly users: common.User[];
@@ -46,7 +46,6 @@ interface Status {
 }
 
 export default class TunerDevice extends EventEmitter {
-
     private _channel: ChannelItem = null;
     private _command: string = null;
     private _process: child_process.ChildProcess = null;
@@ -61,7 +60,7 @@ export default class TunerDevice extends EventEmitter {
     private _exited = false;
     private _closing = false;
 
-    constructor(private _index: number, private _config: config.Tuner) {
+    constructor(private _index: number, private _config: apid.ConfigTunersItem) {
         super();
         this._isRemote = !!this._config.remoteMirakurunHost;
         Event.emit("tuner", "create", this.toJSON());
@@ -72,7 +71,7 @@ export default class TunerDevice extends EventEmitter {
         return this._index;
     }
 
-    get config(): config.Tuner {
+    get config(): apid.ConfigTunersItem {
         return this._config;
     }
 
@@ -131,7 +130,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     getPriority(): number {
-
         let priority = -2;
 
         for (const user of this._users) {
@@ -143,7 +141,7 @@ export default class TunerDevice extends EventEmitter {
         return priority;
     }
 
-    toJSON(): Status {
+    toJSON(): TunerDeviceStatus {
         return {
             index: this._index,
             name: this._config.name,
@@ -208,7 +206,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     endStream(user: User): void {
-
         log.debug("TunerDevice#%d end stream for user `%s` (priority=%d)...", this._index, user.id, user.priority);
 
         user._stream.end();
@@ -228,7 +225,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     async getRemotePrograms(query?: ProgramsQuery): Promise<apid.Program[]> {
-
         if (!this._isRemote) {
             throw new Error(util.format("TunerDevice#%d is not remote device", this._index));
         }
@@ -248,7 +244,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     private _spawn(ch: ChannelItem): void {
-
         log.debug("TunerDevice#%d spawn...", this._index);
 
         if (this._process) {
@@ -270,29 +265,16 @@ export default class TunerDevice extends EventEmitter {
             cmd = this._config.command;
         }
 
-        cmd = cmd.replace("<channel>", ch.channel);
+        cmd = common.replaceCommandTemplate(cmd, {
+            channel: ch.channel,
+            satelite: ch.commandVars?.satellite || "", // deprecated, for backward compatibility
+            space: 0, // default value for backward compatibility
+            ...ch.commandVars
+        });
 
-        if (ch.satellite) {
-            cmd = cmd.replace("<satelite>", ch.satellite); // deprecated
-            cmd = cmd.replace("<satellite>", ch.satellite);
-        }
+        const parsed = common.parseCommandForSpawn(cmd);
 
-        // tslint:disable-next-line:prefer-conditional-expression
-        if (ch.space) {
-            cmd = cmd.replace("<space>", ch.space.toString(10));
-        } else {
-            cmd = cmd.replace("<space>", "0"); // set default value to '0'
-        }
-
-        if (ch.freq !== undefined) {
-            cmd = cmd.replace("<freq>", ch.freq.toString(10));
-        }
-
-        if (ch.polarity) {
-            cmd = cmd.replace("<polarity>", ch.polarity);
-        }
-
-        this._process = child_process.spawn(cmd.split(" ")[0], cmd.split(" ").slice(1));
+        this._process = child_process.spawn(parsed.command, parsed.args);
         this._command = cmd;
         this._channel = ch;
 
@@ -300,14 +282,12 @@ export default class TunerDevice extends EventEmitter {
             const cat = child_process.spawn("cat", [this._config.dvbDevicePath]);
 
             cat.once("error", (err) => {
-
                 log.error("TunerDevice#%d cat process error `%s` (pid=%d)", this._index, err.name, cat.pid);
 
                 this._kill(false);
             });
 
             cat.once("close", (code, signal) => {
-
                 log.debug(
                     "TunerDevice#%d cat process has closed with code=%d by signal `%s` (pid=%d)",
                     this._index, code, signal, cat.pid
@@ -328,7 +308,6 @@ export default class TunerDevice extends EventEmitter {
         this._process.once("exit", () => this._exited = true);
 
         this._process.once("error", (err) => {
-
             log.fatal("TunerDevice#%d process error `%s` (pid=%d)", this._index, err.name, this._process.pid);
 
             ++this._fatalCount;
@@ -343,7 +322,6 @@ export default class TunerDevice extends EventEmitter {
         });
 
         this._process.once("close", (code, signal) => {
-
             log.info(
                 "TunerDevice#%d process has closed with exit code=%d by signal `%s` (pid=%d)",
                 this._index, code, signal, this._process.pid
@@ -365,14 +343,12 @@ export default class TunerDevice extends EventEmitter {
     }
 
     private _streamOnData(chunk: Buffer): void {
-
         for (const user of this._users) {
             user._stream.write(chunk);
         }
     }
 
     private _end(): void {
-
         this._isAvailable = false;
 
         this._stream.removeAllListeners("data");
@@ -388,7 +364,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     private async _kill(close: boolean): Promise<void> {
-
         log.debug("TunerDevice#%d kill...", this._index);
 
         if (!this._process || !this._process.pid) {
@@ -404,7 +379,6 @@ export default class TunerDevice extends EventEmitter {
         this._updated();
 
         await new Promise<void>(resolve => {
-
             this.once("release", resolve);
 
             if (/^dvbv5-zap /.test(this._command) === true) {
@@ -423,7 +397,6 @@ export default class TunerDevice extends EventEmitter {
     }
 
     private _release(): void {
-
         if (this._process) {
             this._process.stderr.removeAllListeners();
             this._process.removeAllListeners();
