@@ -13,6 +13,9 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
+import { promisify } from "util";
+import * as yieldableJSON from "yieldable-json";
+const stringifyAsync = promisify(yieldableJSON.stringifyAsync);
 import * as express from "express";
 
 export interface Error {
@@ -22,7 +25,6 @@ export interface Error {
 }
 
 export function responseError(res: express.Response, code: number, reason?: string): express.Response {
-
     if (reason) {
         res.writeHead(code, reason, {
             "Content-Type": "application/json"
@@ -45,7 +47,6 @@ export function responseError(res: express.Response, code: number, reason?: stri
 }
 
 export function responseStreamErrorHandler(res: express.Response, err: NodeJS.ErrnoException): express.Response {
-
     if (err.message === "no available tuners") {
         return responseError(res, 503, "Tuner Resource Unavailable");
     }
@@ -53,12 +54,11 @@ export function responseStreamErrorHandler(res: express.Response, err: NodeJS.Er
     return responseError(res, 500, err.message);
 }
 
-export function responseJSON(res: express.Response, body: any): express.Response {
-
+export async function responseJSON(res: express.Response, body: any): Promise<express.Response> {
     // this is lighter than res.json()
     res.setHeader("Content-Type", "application/json; charset=utf-8");
     res.status(200);
-    res.end(JSON.stringify(body));
+    res.end(await stringifyAsync(body));
 
     return res;
 }
