@@ -15,15 +15,16 @@
 */
 import { Operation } from "express-openapi";
 import * as api from "../../../api";
+import * as apid from "../../../../../api";
 import _ from "../../../_";
-import { ChannelTypes, ChannelType } from "../../../common";
+import { deepClone, channelTypes } from "../../../common";
 
 export const parameters = [
     {
         in: "path",
         name: "type",
         type: "string",
-        enum: Object.keys(ChannelTypes),
+        enum: channelTypes,
         required: true
     },
     {
@@ -35,24 +36,24 @@ export const parameters = [
 ];
 
 export const get: Operation = (req, res) => {
-
-    const channel = _.channel.get(req.params.type as ChannelType, req.params.channel);
+    const channel = _.channel.get(req.params.type as apid.ChannelType, req.params.channel);
 
     if (channel === null) {
         api.responseError(res, 404);
         return;
     }
 
-    const body: any = channel.toJSON();
+    const body: apid.Channel = deepClone(channel);
 
     body.services = channel.getServices().map(service => ({
         id: service.id,
         serviceId: service.serviceId,
         networkId: service.networkId,
-        name: service.name
+        name: service.name,
+        type: service.type
     }));
 
-    res.json(body);
+    api.responseJSON(res, body);
 };
 
 get.apiDoc = {
