@@ -15,15 +15,16 @@
 */
 import {Operation} from "express-openapi";
 import * as api from "../../../../api";
+import * as apid from "../../../../../../api";
+import { channelTypes } from "../../../../common";
 import _ from "../../../../_";
-import { ChannelType, ChannelTypes } from "../../../../common";
 
 export const parameters = [
     {
         in: "path",
         name: "type",
         type: "string",
-        enum: Object.keys(ChannelTypes),
+        enum: channelTypes,
         required: true
     },
     {
@@ -47,9 +48,8 @@ export const parameters = [
     }
 ];
 
-export const head: Operation = (req, res) => {
-
-    const channel = _.channel.get(req.params.type as ChannelType, req.params.channel);
+export const get: Operation = (req, res) => {
+    const channel = _.channel.get(req.params.type as apid.ChannelType, req.params.channel);
 
     if (channel === null) {
         api.responseError(res, 404);
@@ -58,17 +58,11 @@ export const head: Operation = (req, res) => {
 
     const userId = (req.ip || "unix") + ":" + (req.socket.remotePort || Date.now());
 
-    res.setHeader("Content-Type", "video/MP2T");
-    res.setHeader("X-Mirakurun-Tuner-User-ID", userId);
-    res.status(200).end();
-};
-
-export const get: Operation = (req, res) => {
-
-    const channel = _.channel.get(req.params.type as ChannelType, req.params.channel);
-
-    if (channel === null) {
-        api.responseError(res, 404);
+    // HEAD request support
+    if (req.method === "HEAD") {
+        res.setHeader("Content-Type", "video/MP2T");
+        res.setHeader("X-Mirakurun-Tuner-User-ID", userId);
+        res.status(200).end();
         return;
     }
 
@@ -77,8 +71,6 @@ export const get: Operation = (req, res) => {
 
     (<any> res.socket)._writableState.highWaterMark = Math.max(res.writableHighWaterMark, 1024 * 1024 * 16);
     res.socket.setNoDelay(true);
-
-    const userId = (req.ip || "unix") + ":" + (req.socket.remotePort || Date.now());
 
     channel.getStream({
         id: userId,
@@ -99,31 +91,6 @@ export const get: Operation = (req, res) => {
             res.status(200);
         })
         .catch((err) => api.responseStreamErrorHandler(res, err));
-};
-
-head.apiDoc = {
-    tags: ["channels", "stream"],
-    operationId: "getChannelStream",
-    produces: ["video/MP2T"],
-    responses: {
-        200: {
-            description: "OK",
-            headers: {
-                "X-Mirakurun-Tuner-User-ID": {
-                    type: "string"
-                }
-            }
-        },
-        404: {
-            description: "Not Found"
-        },
-        503: {
-            description: "Tuner Resource Unavailable"
-        },
-        default: {
-            description: "Unexpected Error"
-        }
-    }
 };
 
 get.apiDoc = {
@@ -149,4 +116,12 @@ get.apiDoc = {
             description: "Unexpected Error"
         }
     }
+};
+
+// HEAD request support
+export const head: Operation = (...args) => get(...args);
+
+head.apiDoc = {
+    ...get.apiDoc,
+    operationId: undefined
 };
