@@ -14,12 +14,13 @@
    limitations under the License.
 */
 import { Operation } from "express-openapi";
+import * as api from "../../api";
+import * as apid from "../../../../api";
 import * as config from "../../config";
 
-export const get: Operation = (req, res) => {
-
+export const get: Operation = async (req, res) => {
     res.status(200);
-    res.json(config.loadChannels());
+    api.responseJSON(res, await config.loadChannels() as apid.ConfigChannels);
 };
 
 get.apiDoc = {
@@ -41,14 +42,13 @@ get.apiDoc = {
     }
 };
 
-export const put: Operation = (req, res) => {
+export const put: Operation = async (req, res) => {
+    const channels: apid.ConfigChannels = req.body;
 
-    const channels: config.Channel[] = req.body;
-
-    config.saveChannels(channels);
+    await config.saveChannels(channels);
 
     res.status(200);
-    res.json(channels);
+    api.responseJSON(res, channels);
 };
 
 put.apiDoc = {
