@@ -15,19 +15,59 @@
 */
 import { Operation } from "express-openapi";
 import * as api from "../../../api";
-import * as common from "../../../common";
+import * as apid from "../../../../../api";
 import * as config from "../../../config";
-import * as db from "../../../db";
 import _ from "../../../_";
 
-let isScanning = false;
+/**
+ * Options for a channel scan operation
+ */
+interface ChannelScanOption {
+    type: apid.ChannelType;          // Channel type
+    startCh?: number;                  // Start channel number
+    endCh?: number;                    // End channel number
+    startSubCh?: number;               // Start subchannel number (for BS)
+    endSubCh?: number;                 // End subchannel number (for BS)
+    useSubCh?: boolean;                // Use subchannel style (for BS)
+    scanMode?: apid.ChannelScanMode;               // Scan mode
+    setDisabledOnAdd?: boolean;        // Set disabled flag on newly added channels
+    refresh?: boolean;                 // Refresh existing channel configs
+    channelNameFormat?: string;        // Custom channel name format
+}
 
-const compareOptions = {
+/**
+ * Configuration for a scan operation
+ */
+interface ScanConfig {
+    readonly channels: string[];         // List of channel identifiers to scan
+    readonly scanMode: apid.ChannelScanMode;         // Scan mode to use
+    readonly setDisabledOnAdd: boolean;  // Whether to set disabled on new channels
+}
+
+/**
+ * Service types supported for scanning
+ * 0x01 = デジタルTVサービス
+ * 0x02 = デジタル音声サービス
+ * 0xA1 = 臨時映像サービス
+ * 0xA4 = エンジニアリングサービス
+ * 0xA5 = プロモーション映像サービス
+ * 0xAD = 超高精細度4K専用TVサービス
+ * 0xC0 = データサービス
+ */
+const serviceTypes = [0x01, 0x02, 0xA1, 0xA4, 0xA5, 0xAD, 0xC0];
+
+/**
+ * Options for string comparison when sorting channels
+ */
+const compareOptions: Intl.CollatorOptions = {
     sensitivity: "base" as const,
     numeric: true
 };
 
-const channelOrder = {
+/**
+ * Channel type order for sorting
+ */
+const channelOrder: Record<apid.ChannelType, number> = {
     GR: 1,
     BS: 2,
     CS: 3,
@@ -35,203 +75,276 @@ const channelOrder = {
     BS4K: 5
 };
 
-enum ScanMode {
-    Channel = "Channel",
-    Service = "Service"
-}
+/**
+ * Channel name format templates
+ */
+const CHANNEL_NAME_FORMAT_GR = "{ch}";                // GR channel format
+const CHANNEL_NAME_FORMAT_BS = "{ch}";                // BS channel format
+const CHANNEL_NAME_FORMAT_BS_SUBCH = "BS{ch00}_{subch}"; // BS subchannel format
+const CHANNEL_NAME_FORMAT_CS = "CS{ch}";              // CS channel format
 
-const CHANNEL_NAME_FORMAT_GR = "{ch}";
-const CHANNEL_NAME_FORMAT_BS = "{ch}";
-const CHANNEL_NAME_FORMAT_BS_SUBCH = "BS{ch00}_{subch}";
-const CHANNEL_NAME_FORMAT_CS = "CS{ch}";
+/**
+ * Global flag to track if scan cancellation is requested
+ */
+let isCancellationRequested = false;
 
-interface ChannelScanOption {
-    type: string;
-    startCh?: number;
-    endCh?: number;
-    startSubCh?: number;
-    endSubCh?: number;
-    useSubCh?: boolean;
-    scanMode?: ScanMode;
-    setDisabledOnAdd?: boolean;
-    refresh?: boolean;
-    useNIT?: boolean;
-    channelNameFormat?: string;
-}
+/**
+ * Global object to track current scan status
+ */
+const scanStatus: apid.ChannelScanStatus = {
+    isScanning: false,
+    status: "not_started"
+};
 
-interface ScanConfig {
-    readonly channels: string[];
-    readonly scanMode: ScanMode;
-    readonly setDisabledOnAdd: boolean;
-}
-
+/**
+ * Generate an array of numbers in a range
+ *
+ * @param start The start number of the range (inclusive)
+ * @param end The end number of the range (inclusive)
+ * @returns Array of numbers from start to end
+ */
 function range(start: number, end: number): number[] {
-    return Array.from({ length: (end - start + 1) }, (x, i) => i + start);​​​​​​
+    return Array.from({ length: (end - start + 1) }, (_, i) => i + start);
 }
 
+/**
+ * Format a channel name using placeholders
+ *
+ * @param format The format string with placeholders like {ch}, {ch00}, {subch}
+ * @param ch The channel number
+ * @param subch The optional subchannel number
+ * @returns The formatted channel name
+ *
+ * Supported placeholder formats:
+ * - {ch} - Channel number
+ * - {ch0}, {ch00}, etc. - Channel number with zero padding
+ * - {subch} - Subchannel number
+ * - {subch0}, {subch00}, etc. - Subchannel number with zero padding
+ */
 function formatChannelName(format: string, ch: number, subch?: number): string {
     const values = new Map<string, string>();
-    // {変数名} 形式の文字列をかき集める
+    // Find all placeholders in format {name}
     const placeholders = format.match(/({[\\_a-zA-Z0-9]+})/g);
+
     if (placeholders) {
         for (const placeholder of placeholders) {
-            // {ch[0]*}
+            // Handle {ch} or {ch00} style placeholders
             const chDigits = placeholder.match(/{ch([0]*)}/);
             if (chDigits) {
                 const digits = chDigits[1] ? chDigits[1].length : 1;
-                values.set(placeholder, (ch ? ch : 0).toString(10).padStart(digits, "0"));
+                values.set(placeholder, (ch ?? 0).toString(10).padStart(digits, "0"));
             }
-            // {subch[0]*}
+
+            // Handle {subch} or {subch00} style placeholders
             const subchDigits = placeholder.match(/{subch([0]*)}/);
             if (subchDigits) {
                 const digits = subchDigits[1] ? subchDigits[1].length : 1;
-                values.set(placeholder, (subch ? subch : 0).toString(10).padStart(digits, "0"));
+                values.set(placeholder, (subch ?? 0).toString(10).padStart(digits, "0"));
             }
         }
     }
+
+    // Replace all placeholders with their values
     let formatted = format;
     for (const [key, value] of values) {
         while (formatted.includes(key)) {
             formatted = formatted.replace(key, value);
         }
     }
+
     return formatted;
 }
 
-export function generateScanConfig(option: ChannelScanOption): ScanConfig {
-
-    // delete undefined from option
+/**
+ * Generates a scan configuration for a given channel type and options
+ *
+ * @param option The scan options including channel type, ranges, and formatting
+ * @returns A scan configuration with channels to scan and associated settings
+ */
+export function generateScanConfig(option: ChannelScanOption, usenit:useNIT): ScanConfig | undefined {
+    // Remove undefined properties from options
     Object.keys(option).forEach(key => option[key] === undefined && delete option[key]);
 
-    if (option.type === common.ChannelTypes.GR) {
-        option = {
+    // Handle GR (Ground) channels
+    if (option.type === "GR") {
+        // Set GR-specific defaults
+        const grOptions = {
             startCh: 13,
             endCh: 62,
-            scanMode: ScanMode.Channel,
+            scanMode: "Channel" as const,
             setDisabledOnAdd: false,
             ...option
         };
 
+        // Generate channel list using range and format
+        const channelFormat = grOptions.channelNameFormat || CHANNEL_NAME_FORMAT_GR;
         return {
-            channels: range(option.startCh, option.endCh).map((ch) => formatChannelName(option.channelNameFormat ? option.channelNameFormat : CHANNEL_NAME_FORMAT_GR, ch)),
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            channels: range(grOptions.startCh, grOptions.endCh)
+                .map(ch => formatChannelName(channelFormat, ch)),
+            scanMode: grOptions.scanMode,
+            setDisabledOnAdd: grOptions.setDisabledOnAdd
         };
     }
 
-    option = {
-        scanMode: ScanMode.Service,
+    // Default options for satellite channels (BS/CS)
+    const satelliteOptions = {
+        scanMode: "Service" as const,
         setDisabledOnAdd: true,
         ...option
     };
 
-    if (option.type === common.ChannelTypes.BS4K) {
+    if (option.type === "BS4K") {
         return {
             channels: ["45328"], // default TLV
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            scanMode: satelliteOptions.scanMode,
+            setDisabledOnAdd: satelliteOptions.setDisabledOnAdd
         };
     }
-    if (option.useNIT && option.type === common.ChannelTypes.BS) {
+    if (usenit && option.type === "BS") {
         return {
             channels: ["16625"], // default TS
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            scanMode: satelliteOptions.scanMode,
+            setDisabledOnAdd: satelliteOptions.setDisabledOnAdd
         };
     }
-    if (option.useNIT && option.type === common.ChannelTypes.CS) {
+    if (usenit && option.type === "CS") {
         return {
             channels: ["24608", "28736"], // default TS
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            scanMode: satelliteOptions.scanMode,
+            setDisabledOnAdd: satelliteOptions.setDisabledOnAdd
         };
     }
-    if (option.type === common.ChannelTypes.BS) {
-        if (option.useSubCh) {
-            option = {
+    // Handle BS (Broadcast Satellite) channels
+    if (option.type === "BS") {
+        // Handle subchannel style BS scanning (e.g. BS01_0)
+        if (satelliteOptions.useSubCh) {
+            const bsSubchOptions = {
                 startCh: 1,
                 endCh: 23,
                 startSubCh: 0,
                 endSubCh: 3,
-                ...option
+                ...satelliteOptions
             };
 
+            // Generate cross product of channels and subchannels
             const channels: string[] = [];
-            for (const ch of range(option.startCh, option.endCh)) {
-                for (const subCh of range(option.startSubCh, option.endSubCh)) {
-                    channels.push(formatChannelName(option.channelNameFormat ? option.channelNameFormat : CHANNEL_NAME_FORMAT_BS_SUBCH, ch, subCh));
+            const channelFormat = bsSubchOptions.channelNameFormat || CHANNEL_NAME_FORMAT_BS_SUBCH;
+
+            for (const ch of range(bsSubchOptions.startCh, bsSubchOptions.endCh)) {
+                for (const subCh of range(bsSubchOptions.startSubCh, bsSubchOptions.endSubCh)) {
+                    channels.push(formatChannelName(channelFormat, ch, subCh));
                 }
             }
 
             return {
-                channels: channels,
-                scanMode: option.scanMode,
-                setDisabledOnAdd: option.setDisabledOnAdd
+                channels,
+                scanMode: bsSubchOptions.scanMode,
+                setDisabledOnAdd: bsSubchOptions.setDisabledOnAdd
             };
         }
 
-        option = {
+        // Regular BS channel style (e.g. 101-256)
+        const bsOptions = {
             startCh: 101,
             endCh: 256,
-            ...option
+            ...satelliteOptions
         };
 
+        const channelFormat = bsOptions.channelNameFormat || CHANNEL_NAME_FORMAT_BS;
         return {
-            channels: range(option.startCh, option.endCh).map((ch) => formatChannelName(option.channelNameFormat ? option.channelNameFormat : CHANNEL_NAME_FORMAT_BS, ch)),
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            channels: range(bsOptions.startCh, bsOptions.endCh)
+                .map(ch => formatChannelName(channelFormat, ch)),
+            scanMode: bsOptions.scanMode,
+            setDisabledOnAdd: bsOptions.setDisabledOnAdd
         };
     }
 
-    if (option.type === common.ChannelTypes.CS) {
-        option = {
+    // Handle CS (Communication Satellite) channels
+    if (option.type === "CS") {
+        const csOptions = {
             startCh: 2,
             endCh: 24,
-            ...option
+            ...satelliteOptions
         };
 
+        const channelFormat = csOptions.channelNameFormat || CHANNEL_NAME_FORMAT_CS;
         return {
-            channels: range(option.startCh, option.endCh).map((ch) => formatChannelName(option.channelNameFormat ? option.channelNameFormat : CHANNEL_NAME_FORMAT_CS, ch)),
-            scanMode: option.scanMode,
-            setDisabledOnAdd: option.setDisabledOnAdd
+            channels: range(csOptions.startCh, csOptions.endCh)
+                .map(ch => formatChannelName(channelFormat, ch)),
+            scanMode: csOptions.scanMode,
+            setDisabledOnAdd: csOptions.setDisabledOnAdd
         };
     }
+
+    return undefined;
 }
 
-export function generateChannelItemForService(type: common.ChannelType, channel: string, service: db.Service, setDisabledOnAdd: boolean): config.Channel {
-
-    let name = service.name;
-    name = name.trim();
+/**
+ * Generates a channel item for a specific service
+ *
+ * @param type The channel type (GR, BS, CS, SKY)
+ * @param channel The channel identifier
+ * @param service The service information
+ * @param setDisabledOnAdd Whether to set the channel as disabled initially
+ * @returns A channel configuration object
+ */
+export function generateChannelItemForService(
+    type: apid.ChannelType,
+    channel: string,
+    service: apid.Service,
+    setDisabledOnAdd: boolean
+): apid.ConfigChannelsItem {
+    // Use service name, or fallback to generated name if empty
+    let name = service.name.trim();
     if (name.length === 0) {
         name = `${type}${channel}:${service.serviceId}`;
     }
 
     return {
-        name: name,
-        type: type,
-        channel: channel,
+        name,
+        type,
+        channel,
         serviceId: service.serviceId,
         isDisabled: setDisabledOnAdd
     };
 }
 
-export function generateChannelItemForChannel(type: common.ChannelType, channel: string, services: db.Service[], setDisabledOnAdd: boolean): config.Channel {
-
+/**
+ * Generates a channel item for a channel with multiple services
+ * Finding a common prefix among all service names for the channel name
+ *
+ * @param type The channel type (GR, BS, CS, SKY)
+ * @param channel The channel identifier
+ * @param services Array of services on this channel
+ * @param setDisabledOnAdd Whether to set the channel as disabled initially
+ * @returns A channel configuration object
+ */
+export function generateChannelItemForChannel(
+    type: apid.ChannelType,
+    channel: string,
+    services: apid.Service[],
+    setDisabledOnAdd: boolean,
+    useNIT: boolean
+): apid.ConfigChannelsItem {
+    // Find the common prefix among all service names
     const baseName = services[0].name;
     let matchIndex = baseName.length;
 
+    // Compare each service name with the base name to find the common prefix
     for (let servicesIndex = 1; servicesIndex < services.length; servicesIndex++) {
         const service = services[servicesIndex];
         for (let nameIndex = 0; nameIndex < baseName.length && nameIndex < service.name.length; nameIndex++) {
+            // If characters don't match
             if (baseName[nameIndex] !== service.name[nameIndex]) {
                 if (nameIndex === 0) {
-                    break;
+                    break; // No common prefix at all
                 }
                 if (nameIndex < matchIndex) {
-                    matchIndex = nameIndex;
+                    matchIndex = nameIndex; // Update the match length
                 }
                 break;
             }
+            // If this service name is shorter and is a prefix of the base name
             if (nameIndex + 1 >= service.name.length && service.name.length < matchIndex) {
                 matchIndex = service.name.length;
                 break;
@@ -239,173 +352,580 @@ export function generateChannelItemForChannel(type: common.ChannelType, channel:
         }
     }
 
-    let name = baseName.slice(0, matchIndex);
-    name = name.trim();
-    if (name.length === 0) {
-        name = `${type}${channel}`;
+    // Extract and clean up the common prefix
+    let name = baseName.slice(0, matchIndex).trim();
+    if (useNIT === true) {
+        name = `${type}${channel}`; // Fallback name if no common prefix
+    } else if (name.length === 0) {
+        name = `${type}${channel}`; // Fallback name if no common prefix
     }
 
     return {
-        name: name,
-        type: type,
-        channel: channel,
+        name,
+        type,
+        channel,
         isDisabled: setDisabledOnAdd
     };
 }
 
-export function generateChannelItems(scanMode: ScanMode, type: common.ChannelType, channel: string, services: db.Service[], setDisabledOnAdd: boolean): config.Channel[] {
-
-    if (scanMode === ScanMode.Service) {
-        const channelItems: config.Channel[] = [];
-        for (const service of services) {
-            channelItems.push(generateChannelItemForService(type, channel, service, setDisabledOnAdd));
-        }
-        return channelItems;
+/**
+ * Generates channel items based on scan mode, either per service or per channel
+ *
+ * @param scanMode The scan mode (Service or Channel)
+ * @param type The channel type (GR, BS, CS, SKY)
+ * @param channel The channel identifier
+ * @param services Array of services found on this channel
+ * @param setDisabledOnAdd Whether to set new channels as disabled
+ * @returns Array of channel configuration objects
+ */
+export function generateChannelItems(
+    scanMode: apid.ChannelScanMode,
+    type: apid.ChannelType,
+    channel: string,
+    services: apid.Service[],
+    setDisabledOnAdd: boolean,
+    useNIT: boolean
+): apid.ConfigChannels {
+    // Service mode: create one channel item per service
+    if (scanMode === "Channel") {
+        return [generateChannelItemForChannel(type, channel, services, setDisabledOnAdd, useNIT)];
+    } else if (useNIT === true) {
+        return [generateChannelItemForChannel(type, channel, services, setDisabledOnAdd, useNIT)];
     }
 
-    return [generateChannelItemForChannel(type, channel, services, setDisabledOnAdd)];
+    // Channel mode: create one channel item for all services
+    return services.map(service =>
+        generateChannelItemForService(type, channel, service, setDisabledOnAdd)
+    );
+}
+/**
+ * Base interface for scan status updates
+ */
+interface BaseScanStatusUpdate {
+    channel?: string;
+    type?: apid.ChannelType;
 }
 
-export const put: Operation = async (req, res) => {
+/**
+ * Interface for phase status updates
+ */
+interface PhaseScanStatusUpdate extends BaseScanStatusUpdate {
+    status: apid.ChannelScanPhase;
+    error?: string;
+    saved?: boolean;
+    dryRun?: boolean;
+}
 
-    if (isScanning === true) {
+/**
+ * Interface for step status updates
+ */
+interface StepScanStatusUpdate extends BaseScanStatusUpdate {
+    status: apid.ChannelScanStep;
+    progress?: number;
+    enabled?: boolean;
+    reason?: string;
+    items?: apid.ConfigChannels;
+    count?: number;
+}
+
+/**
+ * Interface for result status updates
+ */
+interface ResultScanStatusUpdate extends BaseScanStatusUpdate {
+    status: apid.ChannelScanResultType;
+    newCount?: number;
+    takeoverCount?: number;
+    result?: apid.ConfigChannels;
+}
+
+/**
+ * Union type for all scan status updates
+ */
+type ScanStatusUpdate = PhaseScanStatusUpdate | StepScanStatusUpdate | ResultScanStatusUpdate;
+
+/**
+ * Executes the channel scanning process
+ *
+ * @param scanConfig Configuration for the scan including channels to scan
+ * @param dryRun If true, don't save changes
+ * @param type Channel type being scanned (GR, BS, CS, SKY)
+ * @param refresh Whether to rescan channels that already exist
+ * @param outputWriter Optional function to write output text during scan
+ * @returns Promise resolving to the final channel list
+ */
+async function runChannelScan(
+    scanConfig: ScanConfig,
+    dryRun: boolean,
+    type: apid.ChannelType,
+    refresh: boolean,
+    outputWriter?: (text: string) => void,
+    skipCh: number[] = [],
+    useNIT: boolean,
+): Promise<apid.ConfigChannels> {
+    try {
+        // Initialize scan data
+        const scanLog: string[] = [];
+        const oldChannelItems = await config.loadChannels();
+        // Filter out channels of the type we're scanning (they'll be replaced)
+        const result: apid.ConfigChannels = oldChannelItems.filter(channel => channel.type !== type);
+        let newCount = 0;
+        let takeoverCount = 0;
+
+        // Initialize global scan status
+        const now = Date.now();
+        Object.assign(scanStatus, {
+            status: "scanning" as const,
+            type,
+            dryRun,
+            progress: 0,
+            currentChannel: "",
+            scanLog,
+            newCount,
+            takeoverCount,
+            result: [],
+            startTime: now,
+            updateTime: now
+        });
+
+        /**
+         * Updates scan phase status
+         * @param update Phase status update
+         * @param textOutput Optional text output for logs
+         */
+        const updatePhaseStatus = (update: PhaseScanStatusUpdate, textOutput?: string): void => {
+            scanStatus.status = update.status;
+            if (update.channel) {
+                scanStatus.currentChannel = update.channel;
+            }
+
+            appendToLog(textOutput);
+        };
+
+        /**
+         * Updates scan step status
+         * @param update Step status update
+         * @param textOutput Optional text output for logs
+         */
+        const updateStepStatus = (update: StepScanStatusUpdate, textOutput?: string): void => {
+            if (update.progress !== undefined) {
+                scanStatus.progress = update.progress;
+            }
+            if (update.channel) {
+                scanStatus.currentChannel = update.channel;
+            }
+            scanStatus.updateTime = Date.now();
+
+            appendToLog(textOutput);
+        };
+
+        /**
+         * Updates scan result status
+         * @param update Result status update
+         * @param textOutput Optional text output for logs
+         */
+        const updateResultStatus = (update: ResultScanStatusUpdate, textOutput?: string): void => {
+            if (update.newCount !== undefined) {
+                scanStatus.newCount = update.newCount;
+            }
+            if (update.takeoverCount !== undefined) {
+                scanStatus.takeoverCount = update.takeoverCount;
+            }
+            if (update.result) {
+                scanStatus.result = update.result;
+            }
+
+            appendToLog(textOutput);
+        };
+
+        /**
+         * Appends text to scan log and writes output
+         * @param textOutput Text to append
+         */
+        const appendToLog = (textOutput?: string): void => {
+            if (textOutput) {
+                const trimmedText = textOutput.trim();
+                scanStatus.scanLog.push(trimmedText);
+                if (outputWriter) {
+                    outputWriter(textOutput);
+                }
+            }
+        };
+
+        // Print dry run notice if applicable
+        if (dryRun) {
+            appendToLog("-- dry run --\n\n");
+        }
+
+
+        // Print scan start message
+        updateStepStatus(
+            { status: "started" as const, type },
+            `channel scanning... (type: "${type}")\n\n`
+        );
+
+        // Process each channel in the scan configuration
+        const totalChannels = scanConfig.channels.length;
+        for (let i = 0; i < totalChannels; i++) {
+            // Check if scanning has been cancelled
+            if (isCancellationRequested) {
+                updatePhaseStatus(
+                    { status: "cancelled" as const },
+                    "Channel scan was cancelled by user request.\n"
+                );
+                break;
+            }
+            const channel = scanConfig.channels[i];
+            const progressPercent = Math.round((i + 1) / totalChannels * 100);
+
+            // Check if this channel should be skipped based on skipCh parameter
+            const channelNumber = parseInt(channel.replace(/[^0-9]/g, ""), 10);
+            if (!isNaN(channelNumber) && skipCh.includes(channelNumber)) {
+                updateStepStatus(
+                    {
+                        status: "skipped" as const,
+                        channel,
+                        reason: "skip_parameter",
+                        progress: progressPercent
+                    },
+                    `channel: "${channel}" (${i + 1}/${totalChannels}) [${progressPercent}%] - skipped by skipCh parameter\n\n`
+                );
+                continue; // Skip to next channel
+            }
+
+            // Update status to show current channel being scanned
+            updateStepStatus(
+                {
+                    status: "scanning_channel" as const,
+                    channel,
+                    progress: progressPercent
+                },
+                `channel: "${channel}" (${i + 1}/${totalChannels}) [${progressPercent}%] ...\n`
+            );
+
+            // Check for existing enabled channels if we're not refreshing
+            if (!refresh) {
+                const existingChannels = oldChannelItems.filter(
+                    item => item.type === type &&
+                           item.channel === channel &&
+                           !item.isDisabled
+                );
+
+                // If there are existing channels, take them over instead of scanning
+                if (existingChannels.length > 0) {
+                    const takeoverInfo = {
+                        status: "takeover" as const,
+                        channel,
+                        count: existingChannels.length,
+                        items: existingChannels
+                    };
+
+                    updateStepStatus(
+                        takeoverInfo,
+                        `-> ${existingChannels.length} existing config found.\n`
+                    );
+
+                    // Add each existing channel to results
+                    for (const channelItem of existingChannels) {
+                        result.push(channelItem);
+                        takeoverCount++;
+                        scanStatus.takeoverCount = takeoverCount;
+                        appendToLog(`-> ${JSON.stringify(channelItem)}\n`);
+                    }
+
+                    updateStepStatus(
+                        {
+                        status: "skipped" as const,
+                            channel,
+                            reason: "existing_config"
+                        },
+                        `# scan has skipped due to the "refresh = false" option because an existing config was found.\n\n`
+                    );
+  
+                    continue; // Skip to next channel
+                }
+            }
+            // Scan the channel for services
+            let services: apid.Service[];
+            let channels: apid.Channel[];
+            try {
+                // Get services from the tuner
+                const r = await _.tuner.getServices(<any> {
+                    type,
+                    channel
+                }, {
+                    id: "Mirakurun:API:channelScan",
+                    priority: 1
+                });
+             services = r.services;
+             channels = r.channels;
+            } catch (error) {
+                // Handle errors (often no signal)
+                const isNoSignalError = /stream has closed before get network/.test(String(error));
+                const errorInfo = {
+                    status: "error" as const,
+                    channel,
+                    reason: isNoSignalError ? "no_signal" : String(error),
+                    progress: progressPercent
+                };
+
+                let errorText = "-> no signal.";
+                if (!isNoSignalError) {
+                    errorText += ` [${error}]`;
+                }
+                errorText += "\n\n";
+
+                updateStepStatus(errorInfo, errorText);
+                continue; // Skip to next channel
+            }
+
+            // Filter services
+            services = services.filter(service => serviceTypes.includes(service.type));
+
+            updateStepStatus(
+                {
+                    status: "services_found" as const,
+                    channel,
+                    count: services.length
+                },
+                `-> ${services.length} services found.\n`
+            );
+
+            // Skip if no services found
+            if (services.length === 0) {
+                appendToLog("\n");
+                continue;
+            }
+
+            // Generate channel items based on scan mode (Service or Channel)
+            const scannedChannelItems = generateChannelItems(
+                scanConfig.scanMode,
+                type,
+                channel,
+                services,
+                scanConfig.setDisabledOnAdd
+            );
+
+            // Add newly scanned items to results
+            const scannedItems: apid.ConfigChannels = [];
+            for (const newChannelItem of scannedChannelItems) {
+                result.push(newChannelItem);
+                newCount++;
+                scanStatus.newCount = newCount;
+                scannedItems.push(newChannelItem);
+                appendToLog(`-> ${JSON.stringify(newChannelItem)}\n`);
+            }
+
+            updateStepStatus(
+                {
+                    status: "channels_found" as const,
+                    channel,
+                    items: scannedItems
+                },
+                `Found ${scannedItems.length} channels for ${channel}\n\n`
+            );
+        }
+
+        // Sort results by type and channel number
+        result.sort((a, b) => {
+            if (a.type === b.type) {
+                return a.channel.localeCompare(b.channel, undefined, compareOptions);
+            } else {
+                return channelOrder[a.type] - channelOrder[b.type];
+            }
+        });
+
+        // Generate and display summary
+        const summaryData = {
+            status: "summary" as const,
+            newCount,
+            takeoverCount,
+            totalTypeCount: newCount + takeoverCount,
+            totalCount: result.length,
+            type
+        };
+
+        updateResultStatus(
+            summaryData,
+            `-> total ${newCount + takeoverCount}/${result.length} (${type}/Any) channels configured.\n\n`
+        );
+
+        updateResultStatus(
+            { status: "summary_new" as const, newCount },
+            `-> new ${newCount} channels found.\n`
+        );
+
+        updateResultStatus(
+            { status: "summary_takeover" as const, takeoverCount },
+            `-> existing ${takeoverCount} channels merged.\n`
+        );
+
+        // If canceled, do not set to Completed
+        if (!isCancellationRequested) {
+            // Save results if not a dry run
+            if (!dryRun) {
+                await config.saveChannels(result);
+                updatePhaseStatus(
+                    { status: "completed" as const, saved: true },
+                    "channel scan has been completed and saved successfully.\n"
+                );
+                updateResultStatus(
+                    { status: "restart_required" as const },
+                    "**RESTART REQUIRED** to apply changes.\n"
+                );
+            } else {
+                updatePhaseStatus(
+                    { status: "completed" as const, dryRun: true },
+                    "channel scan has been completed.\n\n-- dry run --\n"
+                );
+            }
+        }
+
+        // Send final result
+        updateResultStatus(
+            {
+                status: "final_result" as const,
+                result
+            },
+            `Final result: ${result.length} channels\n`
+        );
+
+        return result;
+    } finally {
+        // Always reset scanning and cancellation flags when done
+        scanStatus.isScanning = false;
+        isCancellationRequested = false;
+    }
+}
+
+/**
+ * Get channel scan status - API handler
+ */
+export const get: Operation = async (req, res) => {
+    res.setHeader("Content-Type", "application/json; charset=utf-8");
+
+    // Return current scan status
+    api.responseJSON(res, scanStatus as apid.ChannelScanStatus);
+};
+
+get.apiDoc = {
+    tags: ["config"],
+    summary: "Get Channel Scan Status",
+    description: "Returns the current or last completed scan status and results",
+    operationId: "getChannelScanStatus",
+    produces: [
+        "application/json"
+    ],
+    responses: {
+        200: {
+            description: "OK",
+            schema: {
+                $ref: "#/definitions/ChannelScanStatus"
+            }
+        }
+    }
+};
+
+/**
+ * Initiate a channel scan - API handler
+ */
+export const put: Operation = async (req, res) => {
+    // Check if a scan is already in progress
+    if (scanStatus.isScanning === true) {
         api.responseError(res, 409, "Already Scanning");
         return;
     }
-    isScanning = true;
 
-    req.setTimeout(1000 * 60 * 10); // 10 minites
+    // Set scanning flag
+    scanStatus.isScanning = true;
 
-    const dryRun = req.query.dryRun as any as boolean;
-    const type = req.query.type as common.ChannelType;
-    const refresh = req.query.refresh as any as boolean;
-    const oldChannelItems = config.loadChannels();
-    const result: config.Channel[] = oldChannelItems.filter(channel => channel.type !== type);
+    // Extract query parameters
+    const asyncMode = Boolean(req.query.async);
+    const dryRun = Boolean(req.query.dryRun);
+    const type = req.query.type as apid.ChannelType;
+    const refresh = Boolean(req.query.refresh);
     const useNIT = type === "BS4K" ? !!_.config.server.useStreamId : (type === "BS" || type === "CS" ? !!_.config.server.useTSId : false);
-    let newCount = 0;
-    let takeoverCount = 0;
 
+    // Parse skipCh parameter
+    const skipCh: number[] = req.query?.skipCh as any as number[] || [];
+
+    // Parse channel configuration options
+    const channelOptions: ChannelScanOption = {
+        type,
+        startCh: req.query.minCh ? Number(req.query.minCh) : undefined,
+        endCh: req.query.maxCh ? Number(req.query.maxCh) : undefined,
+        startSubCh: req.query.minSubCh ? Number(req.query.minSubCh) : undefined,
+        endSubCh: req.query.maxSubCh ? Number(req.query.maxSubCh) : undefined,
+        useSubCh: req.query.useSubCh !== undefined ? Boolean(req.query.useSubCh) : undefined,
+        channelNameFormat: req.query.channelNameFormat as string,
+        scanMode: req.query.scanMode as apid.ChannelScanMode,
+        setDisabledOnAdd: req.query.setDisabledOnAdd !== undefined ?
+            Boolean(req.query.setDisabledOnAdd) : undefined,
+        scanMode: req.query.scanMode as apid.ChannelScanMode
+    };
+
+    // Generate scan configuration
+    const scanConfig = generateScanConfig(channelOptions, useNIT);
+
+    // Handle missing scan configuration
+    if (!scanConfig) {
+        scanStatus.isScanning = false;
+        api.responseError(res, 400, "Invalid scan configuration");
+        return;
+    }
+
+    // Handle asynchronous scan mode
+    if (asyncMode) {
+        res.status(202);
+        res.setHeader("Content-Type", "application/json; charset=utf-8");
+        res.json({
+            status: "accepted",
+            message: "Channel scan started in async mode"
+        });
+        res.end();
+
+        // Run scan in background
+        runChannelScan(scanConfig, dryRun, type, refresh, null, skipCh, useNIT)
+            .catch(error => {
+                console.error("Channel scan error:", error);
+                // Error is used only when the scan is stopped
+                scanStatus.status = "error" as const;
+                scanStatus.scanLog.push(`Error: ${String(error)}`);
+                scanStatus.isScanning = false;
+            });
+
+        return;
+    }
+
+    // Synchronous mode - wait for completion
     res.setHeader("Content-Type", "text/plain; charset=utf-8");
     res.status(200);
-    if (dryRun) {
-        res.write("-- dry run --\n\n");
+
+    // Set longer timeout for synchronous scan mode
+    req.setTimeout(1000 * 60 * 30); // 30 minutes
+
+    try {
+        // Stream output directly to response
+        const logTextOutput = (textContent: string): void => {
+            res.write(textContent);
+        };
+
+        // Run scan with output streaming
+        await runChannelScan(scanConfig, dryRun, type, refresh, logTextOutput, skipCh, useNIT);
+        res.end();
+    } catch (error) {
+        console.error("Channel scan error:", error);
+        // Error is used only when the scan is stopped
+        scanStatus.status = "error" as const;
+        scanStatus.scanLog.push(`Error: ${String(error)}`);
+        res.write(`Error during scan: ${error}\n`);
+        res.end();
+        scanStatus.isScanning = false;
     }
-    res.write(`channel scanning... (type: "${type}")\n\n`);
-
-    const scanConfig = generateScanConfig({
-        type: type,
-        startCh: req.query.minCh as any as number,
-        endCh: req.query.maxCh as any as number,
-        startSubCh: req.query.minSubCh as any as number,
-        endSubCh: req.query.maxSubCh as any as number,
-        useSubCh: req.query.useSubCh as any as boolean,
-        channelNameFormat: req.query.channelNameFormat as any as string,
-        scanMode: req.query.scanMode as any as ScanMode,
-        setDisabledOnAdd: req.query.setDisabledOnAdd as any as boolean,
-        useNIT
-    });
-
-    const chLength = scanConfig.channels.length;
-    for (let i = 0; i < chLength; i++) {
-        const channel = scanConfig.channels[i];
-
-        res.write(`channel: "${channel}" (${i + 1}/${chLength}) [${Math.round((i + 1) / chLength * 100)}%] ...\n`);
-
-        if (!refresh) {
-            const takeoverChannelItems = oldChannelItems.filter(chItem => chItem.type === type && chItem.channel === channel && !chItem.isDisabled);
-            if (takeoverChannelItems.length > 0) {
-                res.write(`-> ${takeoverChannelItems.length} existing config found.\n`);
-                for (const channelItem of takeoverChannelItems) {
-                    result.push(channelItem);
-                    ++takeoverCount;
-                    res.write(`-> ${JSON.stringify(channelItem)}\n`);
-                }
-                res.write(`# scan has skipped due to the "refresh = false" option because an existing config was found.\n\n`);
-                continue;
-            }
-        }
-
-        let services: db.Service[];
-        let channels: db.Channel[];
-        try {
-            const r = await _.tuner.getServices(<any> {
-                type: type,
-                channel: channel
-            });
-            services = r.services;
-            channels = r.channels;
-        } catch (e) {
-            res.write("-> no signal.");
-            if (/stream has closed before get network/.test(e) === false) {
-                res.write(` [${e}]`);
-            }
-            res.write("\n\n");
-            continue;
-        }
-
-        if (useNIT) {
-            res.write(`-> ${channels.length} streams found.\n`);
-            for (const c of channels) {
-                if (result.some(x => x.channel === c.channel)) {
-                    continue;
-                }
-                res.write(`-> ${JSON.stringify(c)}\n`);
-                result.push({
-                    name: c.channel,
-                    type,
-                    channel: c.channel
-                });
-                ++newCount;
-            }
-            continue;
-        }
-
-        services = services.filter(service => service.type === 1 || service.type === 173);
-        res.write(`-> ${services.length} services found.\n`);
-
-        if (services.length === 0) {
-            res.write("\n");
-            continue;
-        }
-
-        const scannedChannelItems = generateChannelItems(scanConfig.scanMode, type, channel, services, scanConfig.setDisabledOnAdd);
-        for (const scannedChannelItem of scannedChannelItems) {
-            result.push(scannedChannelItem);
-            ++newCount;
-            res.write(`-> ${JSON.stringify(scannedChannelItem)}\n`);
-        }
-        res.write("\n");
-    }
-
-    result.sort((a, b) => {
-        if (a.type === b.type) {
-            return a.channel.localeCompare(b.channel, undefined, compareOptions);
-        } else {
-            return channelOrder[a.type] - channelOrder[b.type];
-        }
-    });
-
-    res.write(`-> new ${newCount} channels found.\n`);
-    res.write(`-> existing ${takeoverCount} channels merged.\n`);
-    res.write(`-> total ${newCount + takeoverCount}/${result.length} (${type}/Any) channels configured.\n\n`);
-
-    if (!dryRun) {
-        config.saveChannels(result);
-    }
-
-    isScanning = false;
-
-    if (dryRun) {
-        res.write("channel scan has been completed.\n\n");
-        res.write("-- dry run --\n");
-    } else {
-        res.write("channel scan has been completed and saved successfully.\n");
-        res.write("**RESTART REQUIRED** to apply changes.\n");
-    }
-
-    res.end();
 };
 
+/**
+ * API documentation for the channel scan endpoint
+ */
 put.apiDoc = {
     tags: ["config"],
     summary: "Channel Scan",
@@ -434,14 +954,14 @@ About BS Subchannel Style:
             type: "boolean",
             allowEmptyValue: true,
             default: false,
-            description: "dry run. If `true`, the scanned result will not be saved."
+            description: "Dry run mode. If `true`, the scanned result will not be saved to configuration."
         },
         {
             in: "query",
             name: "type",
             type: "string",
-            enum: [common.ChannelTypes.GR, common.ChannelTypes.BS, common.ChannelTypes.CS, common.ChannelTypes.BS4K],
-            default: common.ChannelTypes.GR,
+            enum: ["GR", "BS", "CS", "BS4K"] as apid.ChannelType[],
+            default: "GR",
             description: "Specifies the channel type to scan."
         },
         {
@@ -458,15 +978,26 @@ About BS Subchannel Style:
         },
         {
             in: "query",
+            name: "skipCh",
+            type: "array",
+            items: {
+                type: "integer"
+            },
+            collectionFormat: "csv",
+            description: "Comma-separated list of channel numbers to skip during scanning.\n" +
+                "Example: `skipCh=13,14,15` will skip channels 13, 14, and 15."
+        },
+        {
+            in: "query",
             name: "minSubCh",
             type: "integer",
-            description: "Specifies the minimum number of subchannel numbers to scan. This parameter is only used if the type is `BS` and the bs_subch_style is `true`."
+            description: "Specifies the minimum number of subchannel numbers to scan. This parameter is only used if the type is `BS` and the useSubCh is `true`."
         },
         {
             in: "query",
             name: "maxSubCh",
             type: "integer",
-            description: "Specifies the maximum number of subchannel numbers to scan. This parameter is only used if the type is `BS` and the bs_subch_style is `true`."
+            description: "Specifies the maximum number of subchannel numbers to scan. This parameter is only used if the type is `BS` and the useSubCh is `true`."
         },
         {
             in: "query",
@@ -474,22 +1005,21 @@ About BS Subchannel Style:
             type: "boolean",
             allowEmptyValue: true,
             default: true,
-            description: "Specify true to specify the channel in the subchannel style. Only used for BS scans. (e.g. BS01_0)"
+            description: "Specify true to use the subchannel style for channel numbers. Only used for BS scans. (e.g. BS01_0)"
         },
         {
             in: "query",
             name: "channelNameFormat",
             type: "string",
             allowEmptyValue: true,
-            default: "{ch}",
-            description: "Override format to use for channel name. (e.g. {ch} -> 1, BS{ch00}_{subch} -> BS01_2)"
+            description: "Override format to use for channel name. Supports placeholders like {ch}, {ch00}, {subch}. (e.g. {ch} -> 1, BS{ch00}_{subch} -> BS01_2)"
         },
         {
             in: "query",
             name: "scanMode",
             type: "string",
-            enum: [ScanMode.Channel, ScanMode.Service],
-            description: "To specify the service explictly, use the `Service` mode.\n\n" +
+            enum: ["Channel", "Service"] as apid.ChannelScanMode[],
+            description: "Channel scan mode. Use `Service` mode to create separate entries per service.\n\n" +
                 "_Default value (GR)_: Channel\n" +
                 "_Default value (BS/CS)_: Service"
         },
@@ -498,7 +1028,7 @@ About BS Subchannel Style:
             name: "setDisabledOnAdd",
             type: "boolean",
             allowEmptyValue: true,
-            description: "If `true`, set disable on add channel.\n\n" +
+            description: "If `true`, newly discovered channels will be added in disabled state.\n\n" +
                 "_Default value (GR)_: false\n" +
                 "_Default value (BS/CS)_: true"
         },
@@ -508,16 +1038,125 @@ About BS Subchannel Style:
             type: "boolean",
             allowEmptyValue: true,
             default: false,
-            description: "If `true`, update the existing settings without inheriting them.\n" +
-                "However, non-scanned types of channels will always be inherited."
+            description: "If `true`, update the existing channel configurations without preserving them.\n" +
+                "When false, enabled channels that already exist in the config will be preserved.\n" +
+                "Note: Channels of other types will always be preserved regardless of this setting."
+        },
+        {
+            in: "query",
+            name: "async",
+            type: "boolean",
+            allowEmptyValue: true,
+            default: false,
+            description: "If `true`, the API returns 202 Accepted immediately and performs scan asynchronously.\n" +
+                "Use GET /config/channels/scan to monitor progress and retrieve the result."
         }
     ],
     responses: {
         200: {
-            description: "OK"
+            description: "OK - Synchronous scan completed",
+            schema: {
+                type: "string",
+                description: "Text output of scan process"
+            }
+        },
+        202: {
+            description: "Accepted - Asynchronous scan started",
+            schema: {
+                type: "object",
+                properties: {
+                    status: {
+                        type: "string",
+                        enum: ["accepted"]
+                    },
+                    message: {
+                        type: "string"
+                    }
+                }
+            }
+        },
+        400: {
+            description: "Invalid scan configuration",
+            schema: {
+                $ref: "#/definitions/Error"
+            }
         },
         409: {
             description: "Already Scanning",
+            schema: {
+                $ref: "#/definitions/Error"
+            }
+        },
+        default: {
+            description: "Unexpected Error",
+            schema: {
+                $ref: "#/definitions/Error"
+            }
+        }
+    }
+};
+
+/**
+ * Stop a channel scan in progress - API handler
+ */
+export const del: Operation = async (req, res) => {
+    // Check if a scan is currently in progress
+    if (!scanStatus.isScanning) {
+        api.responseError(res, 404, "No scan in progress");
+        return;
+    }
+
+    // Check if a cancellation is already requested
+    if (isCancellationRequested) {
+        api.responseError(res, 409, "Already Stopping");
+        return;
+    }
+
+    // Set cancellation flag to true to request scan to stop
+    isCancellationRequested = true;
+
+    // Update the scan status
+    scanStatus.status = "cancelled" as const;
+    scanStatus.scanLog.push("Scan cancellation requested by user.");
+
+    // Return success response
+    res.status(206);
+    res.setHeader("Content-Type", "application/json; charset=utf-8");
+    res.json({
+        status: "stopping",
+        message: "Channel scan stop has been requested"
+    });
+};
+
+/**
+ * API documentation for the DELETE channel scan endpoint
+ */
+del.apiDoc = {
+    tags: ["config"],
+    summary: "Stop Channel Scan",
+    description: "Stops a currently running channel scan operation",
+    operationId: "stopChannelScan",
+    produces: [
+        "application/json"
+    ],
+    responses: {
+        206: {
+            description: "Accepted",
+            schema: {
+                type: "object",
+                properties: {
+                    status: {
+                        type: "string",
+                        enum: ["stopping"]
+                    },
+                    message: {
+                        type: "string"
+                    }
+                }
+            }
+        },
+        404: {
+            description: "No scan in progress",
             schema: {
                 $ref: "#/definitions/Error"
             }
