@@ -14,11 +14,10 @@
    limitations under the License.
 */
 import { Operation } from "express-openapi";
-import * as api from "../../api";
-import Event, { EventMessage } from "../../Event";
+import * as apid from "../../../../api";
+import Event from "../../Event";
 
 export const get: Operation = (req, res) => {
-
     res.setHeader("Content-Type", "application/json; charset=utf-8");
     res.status(200);
     res.write("[\n");
@@ -28,8 +27,7 @@ export const get: Operation = (req, res) => {
 
     Event.onEvent(_listener);
 
-    function _listener(message: EventMessage) {
-
+    function _listener(message: apid.Event) {
         if (req.query.resource && req.query.resource !== message.resource) {
             return;
         }
