@@ -41,8 +41,7 @@ export const parameters = [
     }
 ];
 
-export const head: Operation = (req, res) => {
-
+export const get: Operation = (req, res) => {
     const program = _.program.get(req.params.id as any as number);
 
     if (program === null) {
@@ -52,17 +51,11 @@ export const head: Operation = (req, res) => {
 
     const userId = (req.ip || "unix") + ":" + (req.socket.remotePort || Date.now());
 
-    res.setHeader("Content-Type", "video/MP2T");
-    res.setHeader("X-Mirakurun-Tuner-User-ID", userId);
-    res.status(200).end();
-};
-
-export const get: Operation = (req, res) => {
-
-    const program = _.program.get(req.params.id as any as number);
-
-    if (program === null) {
-        api.responseError(res, 404);
+    // HEAD request support
+    if (req.method === "HEAD") {
+        res.setHeader("Content-Type", "video/MP2T");
+        res.setHeader("X-Mirakurun-Tuner-User-ID", userId);
+        res.status(200).end();
         return;
     }
 
@@ -71,8 +64,6 @@ export const get: Operation = (req, res) => {
 
     (<any> res.socket)._writableState.highWaterMark = Math.max(res.writableHighWaterMark, 1024 * 1024 * 16);
     res.socket.setNoDelay(true);
-
-    const userId = (req.ip || "unix") + ":" + (req.socket.remotePort || Date.now());
 
     _.tuner.initProgramStream(program, {
         id: userId,
@@ -95,31 +86,6 @@ export const get: Operation = (req, res) => {
             req.setTimeout(1000 * 60 * 10); // 10 minites
         })
         .catch((err) => api.responseStreamErrorHandler(res, err));
-};
-
-head.apiDoc = {
-    tags: ["programs", "stream"],
-    operationId: "getProgramStream",
-    produces: ["video/MP2T"],
-    responses: {
-        200: {
-            description: "OK",
-            headers: {
-                "X-Mirakurun-Tuner-User-ID": {
-                    type: "string"
-                }
-            }
-        },
-        404: {
-            description: "Not Found"
-        },
-        503: {
-            description: "Tuner Resource Unavailable"
-        },
-        default: {
-            description: "Unexpected Error"
-        }
-    }
 };
 
 get.apiDoc = {
@@ -145,4 +111,12 @@ get.apiDoc = {
             description: "Unexpected Error"
         }
     }
+};
+
+// HEAD request support
+export const head: Operation = (...args) => get(...args);
+
+head.apiDoc = {
+    ...get.apiDoc,
+    operationId: undefined
 };
