@@ -19,10 +19,9 @@ import * as api from "../api";
 import * as apid from "../../../api";
 import _ from "../_";
 import Service from "../Service";
-import { ChannelTypes } from "../common";
+import { channelTypes } from "../common";
 
 export const get: Operation = async (req, res) => {
-
     const serviceItems = [..._.service.items]; // shallow copy
     serviceItems.sort((a, b) => a.getOrder() - b.getOrder());
 
@@ -70,7 +69,7 @@ get.apiDoc = {
             in: "query",
             name: "channel.type",
             type: "string",
-            enum: Object.keys(ChannelTypes),
+            enum: channelTypes,
             required: false
         },
         {
