@@ -15,6 +15,7 @@
 */
 import { Operation } from "express-openapi";
 import * as api from "../../../api";
+import * as apid from "../../../../../api";
 import _ from "../../../_";
 
 export const parameters = [
@@ -28,7 +29,6 @@ export const parameters = [
 ];
 
 export const get: Operation = (req, res) => {
-
     const tuner = _.tuner.get(req.params.index as any as number);
 
     if (tuner === null || Number.isInteger(tuner.pid) === false) {
@@ -36,7 +36,8 @@ export const get: Operation = (req, res) => {
         return;
     }
 
-    api.responseJSON(res, { pid: tuner.pid });
+    const tunerProcess: apid.TunerProcess = { pid: tuner.pid };
+    api.responseJSON(res, tunerProcess);
 };
 
 get.apiDoc = {
@@ -65,8 +66,7 @@ get.apiDoc = {
     }
 };
 
-export const del: Operation = (req, res) => {
-
+export const del: Operation = async (req, res) => {
     const tuner = _.tuner.get(req.params.index as any as number);
 
     if (tuner === null || Number.isInteger(tuner.pid) === false) {
@@ -74,9 +74,14 @@ export const del: Operation = (req, res) => {
         return;
     }
 
-    tuner.kill()
-        .then(() => api.responseJSON(res, {pid: null}))
-        .catch((error: Error) => api.responseError(res, 500, error.message));
+    try {
+        await tuner.kill();
+    } catch (err) {
+        api.responseError(res, 500, err.message);
+        return;
+    }
+
+    res.status(204).end();
 };
 
 del.apiDoc = {
@@ -84,16 +89,8 @@ del.apiDoc = {
     summary: "Kill Tuner Process",
     operationId: "killTunerProcess",
     responses: {
-        200: {
-            description: "OK",
-            schema: {
-                type: "object",
-                properties: {
-                    pid: {
-                        type: "null"
-                    }
-                }
-            }
+        204: {
+            description: "Killed"
         },
         404: {
             description: "Not Found",
