@@ -15,15 +15,14 @@
 */
 import { Operation } from "express-openapi";
 import * as api from "../api";
-import { Version } from "../../../api";
+import * as apid from "../../../api";
 import { getLatestVersion } from "../system";
 const pkg = require("../../../package.json");
 
 export const get: Operation = async (req, res) => {
-
-    const version: Version = {
+    const version: apid.Version = {
         current: pkg.version,
-        latest: getLatestVersion()
+        latest: await getLatestVersion()
     };
 
     api.responseJSON(res, version);
