@@ -16,10 +16,12 @@
 import { execSync } from "child_process";
 import { dirname } from "path";
 import { hostname } from "os";
-import * as fs from "fs";
+import { existsSync, readdirSync } from "fs";
+import { mkdir, copyFile, readFile, writeFile } from "fs/promises";
 import * as yaml from "js-yaml";
 import * as ipnum from "ip-num";
-import * as common from "./common";
+import Queue from "promise-queue";
+import * as apid from "../../api";
 import * as log from "./log";
 
 type Writable<T> = { -readonly [K in keyof T]: T[K] };
@@ -42,102 +44,29 @@ const {
     DISABLE_EIT_PARSING,
     DISABLE_WEB_UI,
     ALLOW_IPV4_CIDR_RANGES,
-    ALLOW_IPV6_CIDR_RANGES
+    ALLOW_IPV6_CIDR_RANGES,
+    ALLOW_ORIGINS,
+    ALLOW_PNA,
+    TSPLAY_ENDPOINT
 } = process.env;
 
 const IS_DOCKER = DOCKER === "YES";
 
-export interface Server {
-    // as Local Server
-    readonly path?: string;
 
-    // as Remote Server
-    readonly port?: number;
+type Server = Readonly<apid.ConfigServer>;
+type Tuner = Readonly<apid.ConfigTunersItem>;
+type Channel = Readonly<apid.ConfigChannelsItem>;
 
-    // hostname
-    readonly hostname?: string;
-
-    /** `true` to disable IPv6 listening */
-    readonly disableIPv6?: boolean;
-
-    readonly logLevel?: log.LogLevel;
-    readonly maxLogHistory?: number;
-
-    readonly maxBufferBytesBeforeReady?: number;
-    readonly eventEndTimeout?: number;
-
-    readonly programGCInterval?: number;
-    readonly epgGatheringInterval?: number;
-    readonly epgRetrievalTime?: number;
-    readonly logoDataInterval?: number;
-    readonly disableEITParsing?: true;
-    readonly disableWebUI?: true;
-    readonly allowIPv4CidrRanges?: string[];
-    readonly allowIPv6CidrRanges?: string[];
-    readonly useTSId?: boolean;
-    readonly useStreamId?: boolean;
-}
-
-export interface Tuner {
-    readonly name: string;
-
-    // GR / BS / CS / SKY / BS4K
-    readonly types: common.ChannelType[];
-
-    // for chardev / dvb
-    readonly command?: string;
-
-    // for dvb
-    readonly dvbDevicePath?: string;
-
-    // for multiplexing w/ remote Mirakurun
-    readonly remoteMirakurunHost?: string;
-    readonly remoteMirakurunPort?: number;
-    readonly remoteMirakurunDecoder?: boolean;
-
-    // decoder
-    readonly decoder?: string;
-    readonly tlvDecoder?: string;
-
-    readonly isDisabled?: boolean;
-}
-
-export interface Channel {
-    readonly name: string;
-
-    // GR / BS / CS / SKY / BS4K
-    readonly type: common.ChannelType;
-
-    // passed to tuning command
-    readonly channel: string;
-    readonly satellite?: string;
-    readonly space?: number;
-    readonly freq?: number;
-    readonly polarity?: "H" | "V";
-
-    // tsmf
-    readonly tsmfRelTs?: number;
-
-    // service id
-    readonly serviceId?: number;
-
-    readonly isDisabled?: boolean;
-
-    /** @deprecated */
-    readonly satelite?: string;
-}
-
-export function loadServer(): Server {
-
+export async function loadServer(): Promise<Server> {
     const path = SERVER_CONFIG_PATH;
 
     // mkdir if not exists
     const dirPath = dirname(path);
-    if (fs.existsSync(dirPath) === false) {
+    if (existsSync(dirPath) === false) {
         log.info("missing directory `%s`", dirPath);
         try {
             log.info("making directory `%s`", dirPath);
-            fs.mkdirSync(dirPath, { recursive: true });
+            await mkdir(dirPath, { recursive: true });
         } catch (e) {
             log.fatal("failed to make directory `%s`", dirPath);
             console.error(e);
@@ -146,19 +75,19 @@ export function loadServer(): Server {
     }
 
     // copy if not exists
-    if (fs.existsSync(path) === false) {
+    if (existsSync(path) === false) {
         log.info("missing server config `%s`", path);
         // copy if not exists
         try {
             log.info("copying default server config to `%s`", path);
-            fs.copyFileSync("config/server.yml", path);
+            await copyFile("config/server.yml", path);
         } catch (e) {
             log.fatal("failed to copy server config to `%s`", path);
             console.error(e);
             process.exit(1);
         }
     }
-    const config: Writable<Server> = load("server", path);
+    const config: Writable<Server> = await load("server", path);
 
     // set default
     if (!config.allowIPv4CidrRanges) {
@@ -166,6 +95,17 @@ export function loadServer(): Server {
     }
     if (!config.allowIPv6CidrRanges) {
         config.allowIPv6CidrRanges = ["fc00::/7"];
+    }
+    if (!config.allowOrigins) {
+        config.allowOrigins = [
+            "https://mirakurun-secure-contexts-api.pages.dev"
+        ];
+    }
+    if (!config.allowPNA) {
+        config.allowPNA = true;
+    }
+    if (!config.tsplayEndpoint) {
+        config.tsplayEndpoint = "https://mirakurun-secure-contexts-api.pages.dev/tsplay/";
     }
 
     if (typeof config.useTSId !== "boolean") {
@@ -187,7 +127,7 @@ export function loadServer(): Server {
             config.hostname = HOSTNAME.trim();
         }
         if (typeof LOG_LEVEL !== "undefined" && /^-?[0123]$/.test(LOG_LEVEL)) {
-            config.logLevel = parseInt(LOG_LEVEL, 10);
+            config.logLevel = parseInt(LOG_LEVEL, 10) as apid.LogLevel;
         }
         if (typeof MAX_LOG_HISTORY !== "undefined" && /^[0-9]+$/.test(MAX_LOG_HISTORY)) {
             config.maxLogHistory = parseInt(MAX_LOG_HISTORY, 10);
@@ -221,6 +161,17 @@ export function loadServer(): Server {
         }
         if (typeof ALLOW_IPV6_CIDR_RANGES !== "undefined" && ALLOW_IPV6_CIDR_RANGES.trim().length > 0) {
             config.allowIPv6CidrRanges = ALLOW_IPV6_CIDR_RANGES.split(",");
+        }
+        if (typeof ALLOW_ORIGINS !== "undefined" && ALLOW_ORIGINS.trim().length > 0) {
+            config.allowOrigins = ALLOW_ORIGINS.split(",");
+        }
+        if (ALLOW_PNA === "true") {
+            config.allowPNA = true;
+        } else if (ALLOW_PNA === "false") {
+            config.allowPNA = false;
+        }
+        if (typeof TSPLAY_ENDPOINT !== "undefined" && TSPLAY_ENDPOINT.trim().length > 0) {
+            config.tsplayEndpoint = TSPLAY_ENDPOINT.trim();
         }
 
         log.info("load server config (merged w/ env): %s", JSON.stringify(config));
@@ -274,17 +225,16 @@ export function saveServer(data: Server): Promise<void> {
     return save("server", SERVER_CONFIG_PATH, data);
 }
 
-export function loadTuners(): Tuner[] {
-
+export async function loadTuners(): Promise<Tuner[]> {
     const path = TUNERS_CONFIG_PATH;
 
     // mkdir if not exists
     const dirPath = dirname(path);
-    if (fs.existsSync(dirPath) === false) {
+    if (existsSync(dirPath) === false) {
         log.info("missing directory `%s`", dirPath);
         try {
             log.info("making directory `%s`", dirPath);
-            fs.mkdirSync(dirPath, { recursive: true });
+            await mkdir(dirPath, { recursive: true });
         } catch (e) {
             log.fatal("failed to make directory `%s`", dirPath);
             console.error(e);
@@ -293,7 +243,7 @@ export function loadTuners(): Tuner[] {
     }
 
     // auto
-    if (fs.existsSync(path) === false) {
+    if (existsSync(path) === false) {
         log.info("missing tuners config `%s`", path);
         log.info("trying to detect tuners...");
         const tuners: Tuner[] = [];
@@ -302,7 +252,7 @@ export function loadTuners(): Tuner[] {
         try {
             execSync("which dvb-fe-tool");
 
-            const adapters = fs.readdirSync("/dev/dvb").filter(name => /^adapter[0-9]+$/.test(name));
+            const adapters = readdirSync("/dev/dvb").filter(name => /^adapter[0-9]+$/.test(name));
             for (let i = 0; i < adapters.length; i++) {
                 log.info("detected DVB device: %s", adapters[i]);
 
@@ -346,7 +296,7 @@ export function loadTuners(): Tuner[] {
         if (tuners.length > 0) {
             try {
                 log.info("writing auto generated tuners config to `%s`", path);
-                fs.writeFileSync(path, yaml.dump(tuners));
+                await saveTuners(tuners);
             } catch (e) {
                 log.fatal("failed to write tuners config to `%s`", path);
                 console.error(e);
@@ -356,11 +306,11 @@ export function loadTuners(): Tuner[] {
     }
 
     // copy if not exists
-    if (fs.existsSync(path) === false) {
+    if (existsSync(path) === false) {
         log.info("missing tuners config `%s`", path);
         try {
             log.info("copying default tuners config to `%s`", path);
-            fs.copyFileSync("config/tuners.yml", path);
+            await copyFile("config/tuners.yml", path);
         } catch (e) {
             log.fatal("failed to copy tuners config to `%s`", path);
             console.error(e);
@@ -375,17 +325,16 @@ export function saveTuners(data: Tuner[]): Promise<void> {
     return save("tuners", TUNERS_CONFIG_PATH, data);
 }
 
-export function loadChannels(): Channel[] {
-
+export async function loadChannels(): Promise<Channel[]> {
     const path = CHANNELS_CONFIG_PATH;
 
     // mkdir if not exists
     const dirPath = dirname(path);
-    if (fs.existsSync(dirPath) === false) {
+    if (existsSync(dirPath) === false) {
         log.info("missing directory `%s`", dirPath);
         try {
             log.info("making directory `%s`", dirPath);
-            fs.mkdirSync(dirPath, { recursive: true });
+            await mkdir(dirPath, { recursive: true });
         } catch (e) {
             log.fatal("failed to make directory `%s`", dirPath);
             console.error(e);
@@ -394,11 +343,11 @@ export function loadChannels(): Channel[] {
     }
 
     // copy if not exists
-    if (fs.existsSync(path) === false) {
+    if (existsSync(path) === false) {
         log.info("missing channels config `%s`", path);
         try {
             log.info("copying default channels config to `%s`", path);
-            fs.copyFileSync("config/channels.yml", path);
+            await copyFile("config/channels.yml", path);
         } catch (e) {
             log.fatal("failed to copy channels config to `%s`", path);
             console.error(e);
@@ -413,32 +362,39 @@ export function saveChannels(data: Channel[]): Promise<void> {
     return save("channels", CHANNELS_CONFIG_PATH, data);
 }
 
-function load(name: "server", path: string): Server;
-function load(name: "tuners", path: string): Tuner[];
-function load(name: "channels", path: string): Channel[];
-function load(name: string, path: string) {
+// use queue because async fs ops is not thread safe
+const configIOQueue = new Queue(1, Infinity);
 
+async function load(name: "server", path: string): Promise<Server>;
+async function load(name: "tuners", path: string): Promise<Tuner[]>;
+async function load(name: "channels", path: string): Promise<Channel[]>;
+async function load(name: "server" | "tuners" | "channels", path: string) {
     log.info("load %s config `%s`", name, path);
 
-    return yaml.load(fs.readFileSync(path, "utf8"));
-}
-
-function save(name: "server", path: string, data: Server): Promise<void>;
-function save(name: "tuners", path: string, data: Tuner[]): Promise<void>;
-function save(name: "channels", path: string, data: Channel[]): Promise<void>;
-function save(name: string, path: string, data: object): Promise<void> {
-
-    log.info("save %s config `%s`", name, path);
-
-    return new Promise<void>((resolve, reject) => {
-
-        fs.writeFile(path, yaml.dump(data), err => {
-
-            if (err) {
-                return reject(err);
-            }
-
-            resolve();
-        });
+    return configIOQueue.add(async () => {
+        return yaml.load(await readFile(path, "utf8"));
     });
 }
+
+async function save(name: "server", path: string, data: Server): Promise<void>;
+async function save(name: "tuners", path: string, data: Tuner[]): Promise<void>;
+async function save(name: "channels", path: string, data: Channel[]): Promise<void>;
+async function save(name: "server" | "tuners" | "channels", path: string, data: object): Promise<void> {
+    log.info("save %s config `%s`", name, path);
+
+    await configIOQueue.add(async () => {
+        await writeFile(path, yaml.dump(data));
+    });
+}
+
+process.on("beforeExit", () => {
+    if (configIOQueue.getQueueLength() + configIOQueue.getPendingLength() === 0) {
+        return;
+    }
+
+    log.warn("configIOQueue is not empty. waiting for completion...");
+
+    setTimeout(() => {
+        log.warn("try to exit again...");
+    }, 100);
+});
