@@ -17,10 +17,10 @@ import * as net from "net";
 import * as http from "http";
 import RPCServer, { Socket } from "jsonrpc2-ws/lib/server";
 import * as log from "./log";
+import * as apid from "../../api";
 import _ from "./_";
 import status from "./status";
 import Event from "./Event";
-import { EventMessage } from "./Event";
 import { event as logEvent } from "./log";
 import { isPermittedHost, isPermittedIPAddress } from "./system";
 import { getStatus } from "./api/status";
@@ -46,7 +46,6 @@ export interface NotifyParams<T> {
  * @experimental
  */
 export function createRPCServer(server: http.Server): RPCServer {
-
     const rpc = new RPCServer({
         pingInterval: 1000 * 30,
         wss: {
@@ -71,13 +70,12 @@ export function createRPCServer(server: http.Server): RPCServer {
 
 const _notifierListeners = new Map<Set<RPCServer>, [Function, Function]>();
 export function initRPCNotifier(rpcs: Set<RPCServer>): void {
-
     const eventsNMDict = {
-        program: new NotifyManager<EventMessage>("events:program", "events", rpcs),
-        service: new NotifyManager<EventMessage>("events:service", "events", rpcs),
-        tuner: new NotifyManager<EventMessage>("events:tuner", "events", rpcs)
+        program: new NotifyManager<apid.Event>("events:program", "events", rpcs),
+        service: new NotifyManager<apid.Event>("events:service", "events", rpcs),
+        tuner: new NotifyManager<apid.Event>("events:tuner", "events", rpcs)
     };
-    function onEventListener(event: EventMessage) {
+    function onEventListener(event: apid.Event) {
         eventsNMDict[event.resource].notify(event);
     }
 
@@ -122,7 +120,6 @@ class NotifyManager<T> {
 }
 
 function serverOnUpgrade(this: RPCServer["wss"], req: http.IncomingMessage, socket: net.Socket, head: Buffer): void {
-
     if (req.socket.remoteAddress && !isPermittedIPAddress(req.socket.remoteAddress)) {
         socket.write("HTTP/1.1 403 Forbidden\r\n\r\n");
         socket.destroy();
