@@ -14,15 +14,17 @@
    limitations under the License.
 */
 import * as os from "os";
+import { promisify } from "util";
+import { exec } from "child_process";
 import { Validator } from "ip-num/Validator";
 import { IPv4, IPv6 } from "ip-num/IPNumber";
 import { IPv4Prefix, IPv6Prefix } from "ip-num/Prefix";
 import { IPv4CidrRange, IPv6CidrRange } from "ip-num/IPRange";
-import { execSync } from "child_process";
 import _ from "./_";
 
-export function getIPv4AddressesForListen(): string[] {
+const asyncExec = promisify(exec);
 
+export function getIPv4AddressesForListen(): string[] {
     const addresses = [];
 
     const interfaces = os.networkInterfaces();
@@ -42,7 +44,6 @@ export function getIPv4AddressesForListen(): string[] {
 }
 
 export function getIPv6AddressesForListen(): string[] {
-
     const addresses = [];
 
     const interfaces = os.networkInterfaces();
@@ -62,7 +63,6 @@ export function getIPv6AddressesForListen(): string[] {
 }
 
 export function isPermittedIPAddress(addr: string): boolean {
-
     const [isIPv4] = Validator.isValidIPv4String(addr);
     if (isIPv4) {
         const ipv4 = new IPv4CidrRange(new IPv4(addr), new IPv4Prefix(32));
@@ -87,7 +87,6 @@ export function isPermittedIPAddress(addr: string): boolean {
 }
 
 export function isPermittedHost(url: string, allowedHostname?: string): boolean {
-
     const u = new URL(url);
 
     if (u.hostname === "localhost" || u.hostname === allowedHostname || isPermittedIPAddress(u.hostname) === true) {
@@ -97,11 +96,9 @@ export function isPermittedHost(url: string, allowedHostname?: string): boolean 
     return false;
 }
 
-export function getLatestVersion(): string {
-
-    const latestVersion = execSync("npm view mirakurun version", {
-        encoding: "utf8"
-    }).trim();
+export async function getLatestVersion(): Promise<string> {
+    const { stdout } = await asyncExec("npm view mirakurun version", { encoding: "utf8" });
+    const latestVersion = stdout.trim();
 
     return latestVersion;
 }
