@@ -88,6 +88,9 @@ export interface ChannelScanOption {
     scanMode?: apid.ChannelScanMode;
     setDisabledOnAdd?: boolean;
     refresh?: boolean;
+    skipCh?: number[];
+    channelNameFormat?: string;
+    async?: boolean;
 }
 
 export class ErrorResponse implements ErrorResponse {
@@ -101,8 +104,7 @@ export class ErrorResponse implements ErrorResponse {
     }
 }
 
-export default class Client {
-
+export class Client {
     basePath = spec.basePath;
     docsPath = "/docs";
     /** positive integer */
@@ -117,17 +119,21 @@ export default class Client {
     private _userAgent = `MirakurunClient/${pkg.version} Node/${process.version} (${process.platform})`;
     private _docs: OpenAPIV2.Document;
 
+    async getDocs() {
+        if (!this._docs) {
+            await this._getDocs();
+        }
+        return this._docs;
+    }
+
     request(method: RequestMethod, path: string, option: RequestOption = {}): Promise<Response>|Promise<ErrorResponse> {
-
         return new Promise((resolve, reject) => {
-
             this._httpRequest(method, path, option).then(
                 res => {
-
                     const ret: Response = {
                         status: res.statusCode,
                         statusText: res.statusMessage,
-                        contentType: res.headers["content-type"].split(";")[0],
+                        contentType: res.headers["content-type"]?.split(";")[0] || "",
                         headers: res.headers,
                         isSuccess: (res.statusCode >= 200 && res.statusCode <= 202)
                     };
@@ -136,7 +142,6 @@ export default class Client {
 
                     res.on("data", chunk => chunks.push(chunk as Buffer));
                     res.on("end", () => {
-
                         const buffer = Buffer.concat(chunks);
 
                         if (ret.contentType === "application/json") {
@@ -155,7 +160,6 @@ export default class Client {
                     });
                 },
                 err => {
-
                     const ret = new ErrorResponse({
                         status: -1,
                         statusText: "Request Failure",
@@ -172,7 +176,6 @@ export default class Client {
     }
 
     async call(operationId: string, param: { [key: string]: any } = {}, option: RequestOption = {}): Promise<any|http.IncomingMessage> {
-
         if (!this._docs) {
             await this._getDocs();
         }
@@ -244,31 +247,26 @@ export default class Client {
     }
 
     async getChannels(query?: ChannelsQuery): Promise<apid.Channel[]> {
-
         const res = await this.call("getChannels", query);
         return res.body as apid.Channel[];
     }
 
     async getChannelsByType(type: apid.ChannelType, query?: ChannelsQuery): Promise<apid.Channel[]> {
-
         const res = await this.call("getChannelsByType", { type, ...query });
         return res.body as apid.Channel[];
     }
 
     async getChannel(type: apid.ChannelType, channel: string): Promise<apid.Channel> {
-
         const res = await this.call("getChannel", { type, channel });
         return res.body as apid.Channel;
     }
 
     async getServicesByChannel(type: apid.ChannelType, channel: string): Promise<apid.Service[]> {
-
         const res = await this.call("getServicesByChannel", { type, channel });
         return res.body as apid.Service[];
     }
 
     async getServiceByChannel(type: apid.ChannelType, channel: string, sid: apid.ServiceId): Promise<apid.Service> {
-
         const res = await this.call("getServiceByChannel", { type, channel, sid });
         return res.body as apid.Service;
     }
@@ -276,7 +274,6 @@ export default class Client {
     async getServiceStreamByChannel(opt: { type: apid.ChannelType, channel: string, sid: apid.ServiceId, decode?: boolean, priority?: number, signal?: AbortSignal }): Promise<http.IncomingMessage>;
     async getServiceStreamByChannel(type: apid.ChannelType, channel: string, sid: apid.ServiceId, decode?: boolean, priority?: number): Promise<http.IncomingMessage>;
     async getServiceStreamByChannel(...args: any[]) {
-
         let type: apid.ChannelType;
         let channel: string;
         let sid: apid.ServiceId;
@@ -311,7 +308,6 @@ export default class Client {
     async getChannelStream(opt: { type: apid.ChannelType, channel: string, decode?: boolean, priority?: number, signal?: AbortSignal }): Promise<http.IncomingMessage>;
     async getChannelStream(type: apid.ChannelType, channel: string, decode?: boolean, priority?: number): Promise<http.IncomingMessage>;
     async getChannelStream(...args: any[]): Promise<http.IncomingMessage> {
-
         let type: apid.ChannelType;
         let channel: string;
         let decode: boolean;
@@ -340,13 +336,11 @@ export default class Client {
     }
 
     async getPrograms(query?: ProgramsQuery): Promise<apid.Program[]> {
-
         const res = await this.call("getPrograms", query);
         return res.body as apid.Program[];
     }
 
     async getProgram(id: apid.ProgramId): Promise<apid.Program> {
-
         const res = await this.call("getProgram", { id });
         return res.body as apid.Program;
     }
@@ -354,7 +348,6 @@ export default class Client {
     async getProgramStream(opt: { id: apid.ProgramId, decode?: boolean, priority?: number, signal?: AbortSignal }): Promise<http.IncomingMessage>;
     async getProgramStream(id: apid.ProgramId, decode?: boolean, priority?: number): Promise<http.IncomingMessage>;
     async getProgramStream(...args: any[]): Promise<http.IncomingMessage> {
-
         let id: apid.ProgramId;
         let decode: boolean;
         let priority: number;
@@ -379,19 +372,16 @@ export default class Client {
     }
 
     async getServices(query?: ServicesQuery): Promise<apid.Service[]> {
-
         const res = await this.call("getServices", query);
         return res.body as apid.Service[];
     }
 
     async getService(id: apid.ServiceItemId): Promise<apid.Service> {
-
         const res = await this.call("getService", { id });
         return res.body as apid.Service;
     }
 
     async getLogoImage(id: apid.ServiceItemId): Promise<Buffer> {
-
         const res = await this.call("getLogoImage", { id });
         return res.body as Buffer;
     }
@@ -399,7 +389,6 @@ export default class Client {
     async getServiceStream(opt: { id: apid.ServiceItemId, decode?: boolean, priority?: number, signal?: AbortSignal }): Promise<http.IncomingMessage>;
     async getServiceStream(id: apid.ServiceItemId, decode?: boolean, priority?: number): Promise<http.IncomingMessage>;
     async getServiceStream(...args: any[]): Promise<http.IncomingMessage> {
-
         let id: apid.ServiceItemId;
         let decode: boolean;
         let priority: number;
@@ -424,117 +413,107 @@ export default class Client {
     }
 
     async getTuners(): Promise<apid.TunerDevice[]> {
-
         const res = await this.call("getTuners");
         return res.body as apid.TunerDevice[];
     }
 
     async getTuner(index: number): Promise<apid.TunerDevice> {
-
         const res = await this.call("getTuner", { index });
         return res.body as apid.TunerDevice;
     }
 
     async getTunerProcess(index: number): Promise<apid.TunerProcess> {
-
         const res = await this.call("getTunerProcess", { index });
         return res.body as apid.TunerProcess;
     }
 
     async killTunerProcess(index: number): Promise<apid.TunerProcess> {
-
         const res = await this.call("killTunerProcess", { index });
         return res.body as apid.TunerProcess;
     }
 
     async getEvents(): Promise<apid.Event[]> {
-
         const res = await this.call("getEvents");
         return res.body as apid.Event[];
     }
 
     async getEventsStream(query?: EventsQuery): Promise<http.IncomingMessage> {
-
         return this.call("getEventsStream", query);
     }
 
     async getChannelsConfig(): Promise<apid.ConfigChannels> {
-
         const res = await this.call("getChannelsConfig");
         return res.body as apid.ConfigChannels;
     }
 
     async updateChannelsConfig(channels: apid.ConfigChannels): Promise<apid.ConfigChannels> {
-
         const res = await this.call("updateChannelsConfig", { body: channels });
         return res.body as apid.ConfigChannels;
     }
 
     async channelScan(option?: ChannelScanOption): Promise<http.IncomingMessage> {
-
         return this.call("channelScan", option);
     }
 
-    async getServerConfig(): Promise<apid.ConfigServer> {
+    async getChannelScanStatus(): Promise<apid.ChannelScanStatus> {
+        const res = await this.call("getChannelScanStatus");
+        return res.body as apid.ChannelScanStatus;
+    }
 
+    async stopChannelScan(): Promise<{ status: string; message: string }> {
+        const res = await this.call("stopChannelScan");
+        return res.body as { status: string; message: string };
+    }
+
+    async getServerConfig(): Promise<apid.ConfigServer> {
         const res = await this.call("getServerConfig");
         return res.body as apid.ConfigServer;
     }
 
     async updateServerConfig(server: apid.ConfigServer): Promise<apid.ConfigServer> {
-
         const res = await this.call("updateServerConfig", { body: server });
         return res.body as apid.ConfigServer;
     }
 
     async getTunersConfig(): Promise<apid.ConfigTuners> {
-
         const res = await this.call("getTunersConfig");
         return res.body as apid.ConfigTuners;
     }
 
     async updateTunersConfig(tuners: apid.ConfigTuners): Promise<apid.ConfigTuners> {
-
         const res = await this.call("updateTunersConfig", { body: tuners });
         return res.body as apid.ConfigTuners;
     }
 
     async getLog(): Promise<string> {
-
         const res = await this.call("getLog");
         return res.body as string;
     }
 
     async getLogStream(): Promise<http.IncomingMessage> {
-
         return this.call("getLogStream");
     }
 
     async checkVersion(): Promise<apid.Version> {
-
         const res = await this.call("checkVersion");
         return res.body as apid.Version;
     }
 
     async updateVersion(force?: boolean): Promise<http.IncomingMessage> {
-
         return this.call("updateVersion", { force });
     }
 
     async getStatus(): Promise<apid.Status> {
-
         const res = await this.call("getStatus");
         return res.body as apid.Status;
     }
 
     async restart(): Promise<{}> {
-
         const res = await this.call("restart");
         return res.body;
     }
 
     private _httpRequest(method: RequestMethod, path: string, option: RequestOption = {}): Promise<http.IncomingMessage> {
-
         const opt: http.RequestOptions = {
             method: method,
             path: this.basePath + path,
@@ -578,9 +557,7 @@ export default class Client {
         }
 
         return new Promise((resolve, reject) => {
-
             const req = http.request(opt, res => {
-
                 if (res.statusCode > 300 && res.statusCode < 400 && res.headers.location) {
                     if (/^\//.test(res.headers.location) === false) {
                         reject(new Error(`Redirecting location "${res.headers.location}" isn't supported.`));
@@ -614,7 +591,6 @@ export default class Client {
     }
 
     private async _requestStream(method: RequestMethod, path: string, option: RequestOption = {}): Promise<http.IncomingMessage> {
-
         const res = await this._httpRequest(method, path, option);
 
         if (res.statusCode >= 200 && res.statusCode <= 202) {
@@ -628,7 +604,6 @@ export default class Client {
     }
 
     private async _getDocs() {
-
         const res = await this.request("GET", this.docsPath);
         if (res.isSuccess !== true) {
             throw new Error(`Failed to get "${this.docsPath}".`);
@@ -636,3 +611,5 @@ export default class Client {
         this._docs = res.body;
     }
 }
+
+export default Client;
