@@ -19,7 +19,7 @@ import * as path from "path";
 
 if (process.argv.length < 3) {
     console.error("Mirakurun EPG Dump Test Program");
-    console.error("Usage: mirakurun-epgdump src.ts dest.json");
+    console.error("Usage: mirakurun-epgdump [-f] src.ts dest.json");
     process.exit(1);
 }
 
@@ -47,11 +47,9 @@ import _ from "./Mirakurun/_";
 import Event from "./Mirakurun/Event";
 import Program from "./Mirakurun/Program";
 import EPG from "./Mirakurun/EPG";
-import * as config from "./Mirakurun/config";
 import * as log from "./Mirakurun/log";
 
 (<any> log).logLevel = log.LogLevel.INFO;
-_.config.server = config.loadServer();
 _.event = new Event();
 _.program = new Program();
 const epg = new EPG();
@@ -65,7 +63,6 @@ const readStream = fs.createReadStream(src);
 
 const transformStream = new stream.Transform({
     transform: function (chunk: Buffer, encoding: string, done: Function) {
-
         bytesRead += chunk.length;
 
         console.log("\u001b[2A");
@@ -75,7 +72,6 @@ const transformStream = new stream.Transform({
         done();
     },
     flush: done => {
-
         console.log("\u001b[2A");
         console.log(`reading - ${bytesRead} of ${size} [${Math.floor(bytesRead / size * 100)}%] (events=${events}) [done]`);
         console.timeEnd("read");
@@ -97,7 +93,6 @@ tsStream.on("eit", (pid, data) => {
 tsStream.resume();
 
 function finalize() {
-
     const programs = Array.from(_.program.itemMap.values());
 
     console.log("programs:", programs.length, "(events)");
