@@ -15,8 +15,10 @@
 */
 import * as React from "react";
 import { useState, useEffect } from "react";
+import { Client as RPCClient } from "jsonrpc2-ws";
 import {
     Text,
+    Link,
     Icon,
     GroupedList,
     DetailsRow,
@@ -31,7 +33,7 @@ import {
     IconButton,
     ColorClassNames
 } from "@fluentui/react";
-import { TunerDevice } from "../../../api";
+import { StreamInfo, TunerDevice } from "../../../api";
 
 interface Item {
     _group: string;
@@ -42,6 +44,9 @@ interface Item {
     controls?: JSX.Element;
     user?: JSX.Element;
     priority?: JSX.Element;
+    ch?: JSX.Element;
+    sid?: JSX.Element;
+    streamInfo?: JSX.Element;
 }
 
 const deviceColumns: IColumn[] = [
@@ -62,15 +67,33 @@ const deviceColumns: IColumn[] = [
 
 const userColumns: IColumn[] = [
     {
+        key: "col-priority",
+        name: "",
+        fieldName: "priority",
+        minWidth: 0
+    },
+    {
         key: "col-user",
         name: "",
         fieldName: "user",
         minWidth: 0
     },
     {
-        key: "col-priority",
+        key: "col-ch",
         name: "",
-        fieldName: "priority",
+        fieldName: "ch",
+        minWidth: 0
+    },
+    {
+        key: "col-sid",
+        name: "",
+        fieldName: "sid",
+        minWidth: 0
+    },
+    {
+        key: "col-streamInfo",
+        name: "",
+        fieldName: "streamInfo",
         minWidth: 0
     }
 ];
@@ -89,12 +112,64 @@ const onRenderCell = (nestingDepth: number, item: Item, itemIndex: number) => {
     />
 };
 
-const TunersManager: React.FC<{ tuners: TunerDevice[] }> = ({ tuners }) => {
+const summarizeStreamInfo = (streamInfo: StreamInfo) => {
+    if (!streamInfo) {
+        return "-";
+    }
 
+    let packets = 0;
+    let drops = 0;
+    for (const pid in streamInfo) {
+        packets += streamInfo[pid].packet;
+        drops += streamInfo[pid].drop;
+    }
+
+    return `Dropped Pkts: ${drops} / ${packets}`;
+};
+
+const isEmptyStreamInfo = (streamInfo: StreamInfo): boolean => {
+    if (!streamInfo) {
+        return true;
+    }
+    return Object.keys(streamInfo).length === 0;
+};
+
+const TunersManager: React.FC<{ tuners: TunerDevice[], rpc: RPCClient }> = ({ tuners, rpc }) => {
     const [killTarget, setKillTarget] = useState<number>(null);
+    const [tunersEx, setTunersEx] = useState<TunerDevice[]>([]);
+    const [streamDetail, setStreamDetail] = useState<{ userId: string; info: StreamInfo }>(null);
+
+    // get streamInfo
+    useEffect(() => {
+        const interval = setInterval(async () => {
+            if (document.hidden) {
+                return;
+            }
+            if (rpc.isConnected() === false) {
+                return;
+            }
+            try {
+                setTunersEx(await rpc.call("getTuners"));
+            } catch (e) {
+                console.warn(e);
+            }
+        }, 1000 * 5);
+
+        return () => clearInterval(interval);
+    }, []);
 
     const items: Item[] = [];
     for (const tuner of tuners) {
+        const tunerEx = tunersEx.find(tunerFull => tunerFull.index === tuner.index);
+        if (tunerEx) {
+            for (const user of tuner.users) {
+                const userEx = tunerEx.users.find(userFull => userFull.id === user.id);
+                if (userEx) {
+                    user.streamInfo = userEx.streamInfo;
+                }
+            }
+        }
+
         const item: Item = {
             _group: `#${tuner.index}: ${tuner.name} (${tuner.types.join(", ")})`,
             _kind: "device",
@@ -146,6 +221,27 @@ const TunersManager: React.FC<{ tuners: TunerDevice[] }> = ({ tuners }) => {
                         <Icon title="Priority" iconName="SortLines" />
                         <Text style={{ marginLeft: 8 }}>{user.priority}</Text>
                     </>
+                ),
+                ch: (
+                    <>
+                        <Icon title="Channel" iconName="TVMonitor" />
+                        <Text style={{ marginLeft: 8 }}>{user.streamSetting.channel.type} / {user.streamSetting.channel.channel}</Text>
+                    </>
+                ),
+                sid: (
+                    <>
+                        <Icon title="Service ID" iconName="Filter" />
+                        <Text style={{ marginLeft: 8 }}>{user.streamSetting.serviceId ? `SID: 0x${user.streamSetting.serviceId.toString(16).toUpperCase()} (${user.streamSetting.serviceId})` : "-"}</Text>
+                    </>
+                ),
+                streamInfo: isEmptyStreamInfo(user.streamInfo) ? undefined : (
+                    <div
+                        style={{ cursor: "pointer" }}
+                        onClick={() => setStreamDetail({ userId: user.id, info: user.streamInfo })}
+                    >
+                        <Icon title="Stream Info" iconName="CubeShape" />
+                        <Link style={{ marginLeft: 8 }}><Text style={{ color: "inherit" }}>{summarizeStreamInfo(user.streamInfo)}</Text></Link>
+                    </div>
                 )
             });
         }
@@ -202,7 +298,63 @@ const TunersManager: React.FC<{ tuners: TunerDevice[] }> = ({ tuners }) => {
                     />
                 </DialogFooter>
             </Dialog>
+
+            <Dialog
+                hidden={!streamDetail}
+                onDismiss={() => setStreamDetail(null)}
+                dialogContentProps={{
+                    type: DialogType.normal,
+                    title: "Stream Info",
+                    subText: `${streamDetail?.userId || ""}`
+                }}
+                minWidth={500}
+            >
+                {streamDetail && <StreamInfoTable userId={streamDetail.userId} tuners={tunersEx} initialInfo={streamDetail.info} />}
+                <DialogFooter>
+                    <DefaultButton
+                        text="Close"
+                        onClick={() => setStreamDetail(null)}
+                    />
+                </DialogFooter>
+            </Dialog>
         </>
+    );
+};
+
+const StreamInfoTable: React.FC<{
+    userId: string;
+    tuners: TunerDevice[];
+    initialInfo: StreamInfo;
+}> = ({ userId, tuners, initialInfo }) => {
+    // Find current streamInfo from tuners
+    let currentInfo = initialInfo;
+    for (const tuner of tuners) {
+        const user = tuner.users.find(u => u.id === userId);
+        if (user) {
+            currentInfo = user.streamInfo;
+            break;
+        }
+    }
+
+    return (
+        <table style={{ width: "100%", borderCollapse: "collapse" }}>
+            <thead>
+                <tr>
+                    <th style={{ padding: "8px", textAlign: "left", borderBottom: "1px solid #ccc" }}>PID</th>
+                    <th style={{ padding: "8px", textAlign: "right", borderBottom: "1px solid #ccc" }}>Packets</th>
+                    <th style={{ padding: "8px", textAlign: "right", borderBottom: "1px solid #ccc" }}>Drops</th>
+                </tr>
+            </thead>
+            <tbody>
+                {Object.entries(currentInfo || {}).map(([pid, data]) => (
+                    <tr key={pid} style={{ borderBottom: "1px solid rgb(52, 54, 63)"}}>
+                        <td style={{ padding: "4px 8px", textAlign: "left" }}>{pid}</td>
+                        <td style={{ padding: "4px 8px", textAlign: "right" }}>{data.packet}</td>
+                        <td style={{ padding: "4px 8px", textAlign: "right" }}>{data.drop}</td>
+                    </tr>
+                ))}
+            </tbody>
+        </table>
     );
 };
 
