@@ -1,9 +1,9 @@
-var assert = require('assert');
+const { describe, it } = require("node:test");
+const assert = require("assert");
 
 const scan = require("../lib/Mirakurun/api/config/channels/scan");
 
-describe("Mirakurun/api/config/channel/scan : generateScanConfig", () => {
-
+describe("[scan.spec] /api/config/channel/scan : generateScanConfig", () => {
     it("GR: Type only", () => {
         const config = scan.generateScanConfig({
             type: "GR"
@@ -504,8 +504,7 @@ describe("Mirakurun/api/config/channel/scan : generateScanConfig", () => {
     });
 });
 
-describe("Mirakurun/api/config/channel/scan : generateChannelItemForService", () => {
-
+describe("[scan.spec] /api/config/channel/scan : generateChannelItemForService", () => {
     it("GR Regular case", () => {
         const type = "GR";
         const ch = "10";
@@ -759,8 +758,7 @@ describe("Mirakurun/api/config/channel/scan : generateChannelItemForService", ()
     });
 });
 
-describe("Mirakurun/api/config/channel/scan : generateChannelItemForChannel (single service)", () => {
-
+describe("[scan.spec] /api/config/channel/scan : generateChannelItemForChannel (single service)", () => {
     it("GR Regular case", () => {
         const type = "GR";
         const ch = "10";
@@ -1000,8 +998,7 @@ describe("Mirakurun/api/config/channel/scan : generateChannelItemForChannel (sin
     });
 });
 
-describe("Mirakurun/api/config/channel/scan : generateChannelItemForChannel (multiple service)", () => {
-
+describe("[scan.spec] /api/config/channel/scan : generateChannelItemForChannel (multiple service)", () => {
     it("Name Summary regular case.1", () => {
         const type = "GR";
         const ch = "10";
@@ -1571,8 +1568,7 @@ describe("Mirakurun/api/config/channel/scan : generateChannelItemForChannel (mul
     });
 });
 
-describe("Mirakurun/api/config/channel/scan : generateChannelItems", () => {
-
+describe("[scan.spec] /api/config/channel/scan : generateChannelItems", () => {
     it("Service mode case.1", () => {
         const mode = "Service";
         const type = "GR";
